[roll] Roll fuchsia [virtualization][linux_runner] Add LinuxManager FIDL.
This exposes a LinuxManager protocol from the linux runner
that can be used to launch a linux guest and monitor the
startup process.

Vsh has been updated to take advantage of this and the
vsh-terminal component can as a result be used to launch
and enter a Linux container.

Test: vsh -c
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620981
Original-Revision: 83dbcb73d26a2e556af9cc82ae5ac43edf81778c
GitOrigin-RevId: 14cebfb6174ba54477f2f1084e70550c9911ef9b
Change-Id: Ib0edf02a39f3abf208812dca2eb2053c44792c3e