[roll] Roll infra/recipes [fuchsia_proto] Add use_ffx_experimental field.
This would go in after http://fxr/617022 which uses the flag to
determine whether to run experimental ffx features. This is useful for
when we want to roll out some ffx features but not others that are still
undergoing experimentation.

Original-Bug: 84869
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/620561
Original-Revision: 49bb8c2be8601c865960bd41dbd4623f52e4a034
GitOrigin-RevId: 9967185a14ca9006ea9a8a20869ff797ad904827
Change-Id: I17e91bf52eb209eddf510e82ef53cb47b9add917