[roll] Roll fuchsia [assembly] Fix assembly inputs collection
This change does two things
1. includes the package manifests as inputs
2. drop all the inputs into "built/artifacts/"

We are putting all the artifacts into "built/artifacts/" so that when a
path has the "../../" prefix it will end up in the root. This makes
OOT integration much easier, because paths in the product config, board
config, and package manifests do not need to be modified.

Original-Bug: 85370
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621664
Original-Revision: c2cae17978ae3887667922acbf5a807f47223005
GitOrigin-RevId: 3dabaa22d1f49e69033b85b51aa277ad2f9621f1
Change-Id: I877607211d72646a242d4931e2f4e151978e5622