[roll] Roll fuchsia [ui-input][input-pipeline] add mouse_binding::MousePhase
Add the moue_binding::MousePhase enum, which represents the phases
which are relevant to a mouse device.

This eliminates the need for upper layers (e.g. ClickDragHandler)
to validate that the phase of a MouseEvent is valid for a mouse.

Original-Bug: 90516
Test: unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621281
Original-Revision: 70e73cdeac4edc83024037e9162d35e02eb90bf1
GitOrigin-RevId: 422e65722d5c27541aa6d89084901435a55819a0
Change-Id: I54ec8305711bb7032ded64e92dc6896a6470bf68