[roll] Roll fuchsia [netstack3] Remove try_unit
This macro is a shorthand that saves a few lines at the expense of
readability; removing it also revealed a number of inconsistent usages
with or without Option::ok_or_else.

This macro also bakes-in Result::ok, which is contrary to networking
rust patterns.

Remove `impl From<Ipv6SourceAddr> for Option<UnicastAddr<Ipv6Addr>>` and
`Ipv6SourceAddr::to_option` as the former is unused and the latter has a
single remaining caller which is trivially removed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620784
Original-Revision: 7179d9898eb188d5cf3ef117ec688f09a0631d35
GitOrigin-RevId: 24e32614a568aa6348c4c1fe5259f0c1f10b2e3a
Change-Id: I2fd2947bf246515becf22c349d1613607350dc0c