[roll] Roll fuchsia [assembly] Allow the blob.manifest to be located outside CWD
The blob.manifest contains paths to blobs relative to the blob.manifest
itself. Previously, the assembly code required that the blob.manifest
was located underneath the CWD hierarchy of files. In order to support
out-of-tree workflows, this does not work, therefore we are making it
more flexible.

Original-Bug: 85370
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621901
Original-Revision: 78776836efd8beb822f10e27904d37c12b54f82a
GitOrigin-RevId: 243d598629022da1201e0127c56e8cdf53c5bd2d
Change-Id: I91e55f4e770f358cb2e4818c05e0ac11196accb4