[roll] Roll fuchsia [netstack] Shorten autogenerated test case name
ResultDB sets a limit of 512 characters for a test case name. When
no base name is given to the test_case macro, it autogenerates a name
using the remaining contents. In this case, the autogenerated name is
over 500 characters and is close to exceeding the limit of what
ResultDB will accept.
This change specifies a base name to greatly shorten the autogenerated
test case name.

Testing: Manually run fx test netstack-inspect-integration-test with
and without this change and verify the same number of tests are run.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621902
Original-Revision: 268bb8f6838ba85d6bd73a9d02c572c15c1c016b
GitOrigin-RevId: 8e7da657fe1b89464f23859f85fd37f0f5fba3e2
Change-Id: I9256e2a2695b161055bc965602abefa5613d2084