[roll] Roll fuchsia [kernel][vm] Make over unmapping explicit
Extends ArchVmAspace::Unmap to take a parameter that controls whether
or not unmap can operate on a larger range than requested. A larger
range might be unmapped if a large page needs to be split, and this
can fail if there's no memory for a page table. For the kernel aspace
over unmapping is not tolerable, since the kernel cannot fault the
the pages back in later, unlike a user address space.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/615948
Original-Revision: 38f6ea745ce7841d8b10a702f9c4715c5de3aca5
GitOrigin-RevId: 71fac6070a72504c1a88a1805fa8605ed989a6c3
Change-Id: Icf1ee1fe5d7c3fb6f6df1d28e59723e800cfe71d