[roll] Roll fuchsia [realm_builder][rust] implement client side of new API
Introduces a new implementation for the RealmBuilder struct that is
based on the new fuchsia.component.test/Realm and /Builder protocols,
instead of the old fuchsia.component.test/RealmBuilder protocol.

This new implementation is in a new module named `soft_migration`. After
this lands clients will migrate to the new definitions under
`soft_migration`, and once this is complete the definitions for the new
implementation will be moved to the top-level of this crate (and
reexported under the old soft_migration module until clients migrate
back to the top-level definitions).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/613645
Original-Revision: aabf6ee6615d02b36e751779059310073dbd98b1
GitOrigin-RevId: 65aba5964bc4f8cb60f9ed055cc44c01ec6ac11e
Change-Id: I93c3b34f7e613b9adfebce5c54533f27d3b5b5ff