[roll] Roll fuchsia [virtualization][linux_runner] Switch to bullseye image
Bullseye is now the latest stable release of debian.

This also avoids using the debian version in the
container name as that requires it to be updated
each time we switch to a new version of debian. Also
going to be confusing when upgrading containers.

Fixes: 90525
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621381
Original-Revision: 10f153d281453ba7b91ad655ab2878af159f8030
GitOrigin-RevId: f3dabcafd3dc553610af5e0cb50e2327229382ef
Change-Id: I6ef5ebe4291c38835e2f484d6f68c07439563d08