[roll] Roll fuchsia [component][fuzzing] Fix idle_ race in Controller
This CL fixes a data race between the FIDL dispatcher and the
Controller's worker thread. Only one workflow is allowed to running at
any one time for a given fuzzer, and this was checked using an `idle_`
flag. Previously a small window existed between the execution of the
callback for a FIDL call and the idle flag being reset to true. A client
could receive the FIDL response and immediately make a subsequent valid
request, only to be told a workflow was in progress.

This race was previously observed, and incorrectly attributed to being
caused by the request being received before the worker loop was entered.
The correct fix is to have the idle flag set initially, and reset under
the lock at the same time that the callback is invoked.

MULTIPLY: .*component-fuzzing-common-tests.* (fuchsia): 30

Original-Bug: 86277
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/622083
Original-Revision: 2615bdd631199684efee65867309b01e9b262dbd
GitOrigin-RevId: ddcef6989bb275d52680bcd0c921943a035b5b70
Change-Id: Ie5d1d4849f39f01afa58c90c46ae5322a457b5b1