[roll] Roll infra/recipes [coverage] Upload absolute coverage in smaller chunks.
The step was failing due to a timeout, so I increased the timeout as
well, but it fails for a different reason (see bug). This workaround
breaks up the coverage data to upload in smaller chunks.
Also fail the build if the step fails.

Original-Bug: 89547, 90510
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/622961
Original-Revision: 21861b778ca8698307a6478bfabca1626e258211
GitOrigin-RevId: 165bcd46b71e5f31e1365a39108ebb7b29693575
Change-Id: Id80662a7a76106dbbf9fb8b08384477e152b81f0