[roll] Roll fuchsia [vfs] Remove abstract Inspector class
Was used for allowing blocks to be read from a filesystem instance, but
replaced with a callback to simplify things and reduce confusion with
the Inspect library.

Remove InspectableMinfs base class as it is no longer required, and
updated tests accordingly.

Test: fx test minfs-unit-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621628
Original-Revision: ff64714176a675d790deb5e76b2afc22ebf15a35
GitOrigin-RevId: a197c1fc5214315877cb20835c0cddd7a5c13c3c
Change-Id: Ic4c339a6d10b53d700babe1116bfc17be7ac39b0