[roll] Roll fuchsia [botanist] Fix race between runAgainstTarget() and target.Wait().
When runAgainstTarget() completes, it shuts down the target. However,
the target calls Wait() in another goroutine to see if it gets any other
errors so that it can exit early. If the target gets shut down properly,
but we receive the target.Wait() error before getting the
runAgainstTarget() error, then we could falsely fail the task.
Cancelling the errgroup context once runAgainstTarget() is finished and
checking for the context error in the other goroutines should prioritize
the error we get from runAgainstTarget() in the case there is a race.

Original-Bug: 90679
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/623201
Original-Revision: 362e5bc40a88979824e2bb80cead19659dd8b6fc
GitOrigin-RevId: b82960830445c5f363042eb82296cad34007be21
Change-Id: If935f1548ffa64c3a816a77fdefa72faed6a125f