[roll] Roll fuchsia [sshutil] Fix noisy logs
"error closing keepalive session" and "closing ssh session: EOF" would
previously be printed in nearly every task, whether it passed or failed,
because session.Close() returns an error if the connection has already
intentionally been closed by testrunner or botanist. So if Close() fails
because the session was already closed, it means we're shutting down
normally so we should ignore the error, causing it to not be logged.

Original-Bug: 90137
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/623222
Original-Revision: 4cd229d4afb0e1141c1b9d903808ea3132495744
GitOrigin-RevId: 2f4605b17ca3e82f07e6caafc91654e0a54bbaac
Change-Id: Ia03330fdef46c8018bc012e6c1534747688f6080