[roll] Roll fuchsia [audio_core][test] Fix flaky audio clock test case
Clock tests are written to skip the actual synchronization
checks if any kind of underflow or overflow occur, but also
to automatically fail to bring attention to those underflows.
The cause of these underflows continues to be investigated
and addressed (tracked by fxb/80003), so in the meantime we
only automatically fail on Renderer underflows, based on the
assumption that the test cases submit the entire source
signal before starting playback.

Unfortunately this is not the case for these test cases, so
all underflow checks have been disabled until fxb/80003 is
fixed.

Multiply: audio-clock-pipeline-test (fuchsia)
Original-Fixed: fxb/79916

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/623289
Original-Revision: 5ded1810dea792a6125ad4476540e884ccfe4c09
GitOrigin-RevId: 439b8d8e890c01e6086ddd986188efd579c61baa
Change-Id: Iae007dd88e8051cf7e7e7ac868e9d338cb948561