[roll] Roll fuchsia [pretty] Update unit k -> K
The NIST(National Institute for Standard and Technology) defines:
kibi for kilobyte with KiB as symbol.
https://physics.nist.gov/cuu/Units/binary.html

The kB can be interpreted as Kilobytes or kibi (1000 or 1024).

The library is consistent with powers of 2, which is why it should
output units with an uppercase K. Even the cpp SizeUnit is KiB.

This change also makes the unit consistent with the rest.

Interesting stack overflow discussion regarding the subject, includes
references to standard as well:
https://superuser.com/questions/938234/size-of-files-in-windows-is-it-kb-or-kb

No functional changes.

Test: CQ
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/623282
Original-Revision: cb1162fa65cfd23197a87d06453f2da6da72dd7a
GitOrigin-RevId: cfc8c4900fde12c394957869f16fa2ea06ef03d8
Change-Id: I365e0761ebef4040a078ccd71ba7a857faadb712