[roll] Roll fuchsia [debugger] Improve PLT stepping.
Previously stepping through PLT entries triggered a simple heuristic
about unsymbolized code that caused it to be single-stepped until we got
to symbolized code. The stepper would then re-evaluate the situation
when the single-stepping is complete.

This had two problems:

 - If the destination symbol hasn't been resolved yet, this will
   single-step through all the dynamic loader code to do this which
   takes an extremely long time. Fuchsia currently resolves all symbols
   up-front so this doesn't apply, but it will in the future.

 - If the destination of the call is not symbolized, this will end up
   single-stepping through the entire function which can have
   catastrophic performance problems. Especially out-of-tree, some
   developers don't always have symbols set up. It's not practically
   possible to even step over unsymbolized cross-library calls.

This patch adds a dedicated thread controller to step through PLT
entries. It resolves the destination of the PLT jump and sets a
breakpoint at that location. This has the effect of skipping any loader
code without having to single step it.

This will also enable us to change our unsymbolized code handling policy
to solve the second problem (though it does not do this yet).

Test updates

 - The new "function" stepping unit tests is moved from the "step into"
   test and enhanced with the more specific PLT behavior.

 - The "SharedLibThunkTest" was removed because it's a now-obsolete
   test of the same behavior.

 - This change makes the "until" thread controller more picky about
   actually matching its breakpoint to the exception it matches. Some
   tests required updating to supply the correct state.

Original-Bug: 60243, 80275
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/619151
Original-Revision: 4df470c0adcb6980c97802622441adac1dceb906
GitOrigin-RevId: c722a0462194fd3678ecf4bc023c7eaab8d8033f
Change-Id: I57d47932cfefe82565cbf2b779c7185fe28a3da5