[roll] Roll fuchsia [zircon][time] Remove ZX_CLOCK_UTC.
This enumeration was defined mainly for use in the zx_clock_get syscall
that was deleted ~6 months ago in accordance with RFC-0008. We had a
few additional usages, particularly of the derived zx::time_utc type,
but have steadily burnt those down over the last couple of months and
we're ready to remove the constant and with it the last vestige of UTC
as a kernel responsibility!

Original-Bug: 86976
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/614036
Original-Revision: 6bd790832163e01447b437c80638015fe13b289c
GitOrigin-RevId: e3faf64d3250df247ae6d16cb3ab3da7622cfa75
Change-Id: I551550b8434d0ee68258b84b8efdd70e248aaca9