[roll] Roll fuchsia [assembly] Disallow absolute paths in assembly inputs
In order for assembly inputs to be portable out-of-tree, absolute paths
cannot be used. In the prebuilt_packages, absolute paths were used to
create the PackageManifest (using pm expand). This CL changes the
absolute path to a relative path, and raises an Exception if any other
absolute paths are found.

Original-Bug: 85370
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/622241
Original-Revision: d314e01884350ac107e30644091eade3a7e28f53
GitOrigin-RevId: 95599c3e0f32648bf30cd4564e37f68e4e375d02
Change-Id: I0b8f8e0402cddb5113bee297e9b2a08599a3b08b