[roll] Roll fuchsia [netstack] Remove a panic from address removal
Removing an address used to do lookups to ensure that the interface
and the address exists, and then panic if `stack.RemoveAddress` returns
an error that indicates that the interface no longer exists.

Don't do the lookups, since it's possible for the interface to be
removed between doing the lookup and calling `stack.RemoveAddress`.

Add a regression integration test which removes addresses and the
interface to which the addresses belong at the same time.

Original-Fixed: 90396
Test: fx test --test-filter 'remove_interface_and_address_*' \
              netstack-fidl-integration-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/619961
Original-Revision: 7779a3d4269a481f86b1152939e37654bf873a06
GitOrigin-RevId: 0a9d28f8b20e3fa32c17be8f08a99e4a09088ec4
Change-Id: I4bf5029dc4bf5b27e45292218d8c094328902666