[roll] Roll fuchsia [llcpp] Remove Transport parameter from ClientEnd and ServerEnd
This CL removes the transport parameter from ClientEnd and ServerEnd and
instead retrieves the transport from a using declaration on the protocol
marker. A field on the protocol marker was used for this rather than a new
trait because the transport needs to be accessible in generated wire type
files.

Additionally, this CL:
- Removes dependencies on event and sync client template logic for the
driver transport. This reduces the amount of existing channel-specific
code that needs to change.
- Adds a dependency on the driver_runtime for LLCPP libraries in
fidl_cpp.gni. This is temporary and was necessary to resolve dependency
issues (wire types depend on ClientEnd/ServerEnd which depends on
the driver transport).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620923
Original-Revision: 4539638643ec74b4430e875dc3ba22a5ae8d720a
GitOrigin-RevId: fdd4e13ead8ed3807a9cb7a1557e45826b156f11
Change-Id: I62e81ee9341670db28213dcfefb77fd010020cfc