[roll] Roll fuchsia [component_manager] Destroy event stream task on component purge
In EventStreamProvider, the task that was serving the event stream over
FIDL was being detached. This led to an issue where if the component
was purged, the serving task would continue running.

This resulted in a handle + memory leak in CM.

The task has now been tied to the component lifecycle by adding it to
EventStreamAttachment.

EventStreamAtttachments for a component are cleaned up on Purge.

Original-Bug: 80731
Test: fx set core.x64 --with //src/sys/component_manager:tests
      fx test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/623030
Original-Revision: d8c8305035489f52eb4854989b15fbf3f9ea8574
GitOrigin-RevId: b7d8cc57a603d458c0d660073f97ce59492a7d23
Change-Id: I4371d8af87b9a17276d9b030f4783100600be34a