[roll] Roll infra/recipes Stop running steps after cancellation
These are all places where we catch a StepFailure and then try to run
more steps. However, if the StepFailure was raised as a result of the
build being cancelled, trying to run any further steps will raise
InfraFailures that cause the build to turn purple instead of blue.

For example:
- https://ci.chromium.org/b/8827207562702022737
- https://ci.chromium.org/b/8827498073560797329

As a workaround, ensure we stop running steps when we receive a
StepFailure as a result of the build being cancelled.

This requires some unfortunate copy-pasting to handle the autocorrelator
case, but I'm working on a followup CL to deduplicate that code.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/623622
Original-Revision: ed89e0e52b1b8fa9c46d5a5253115d73b92969ec
GitOrigin-RevId: c60c17dfbc36e0a4a9a65c4f408835d62eef3afc
Change-Id: I17971d630446f18d76e030b681d8c9750a66b6d5