[roll] Roll fuchsia [blobfs] add new component runner
Add a new Runner class to blobfs for running in a component context.
Running as a component requires us to parameterize blobfs in a slightly
awkward way. When it's first launched, it serves a new fidl protocol,
fuchsia.fs.startup.Startup, which takes a block device and launch
arguments for launching the filesystem, as well as providing fsck- and
mkfs-like functions. Once the launch function is called, the filesystem
is configured, which represents a second stage of startup procedure.

This also adds a new blobfs-component binary which launches blobfs using
this new runner, and a component-level integration test for it.

Original-Bug: 82896
Test: fx test blobfs-component-runner-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620886
Original-Revision: 3a0e065b39eb1a8d4efb5afdae423b7d0f003b47
GitOrigin-RevId: cc0296e87e06c2cfc031c4fb999883ed77081410
Change-Id: I7e44d3d444eb1463a029a59f57c2efacca25328a