[roll] Roll fuchsia [debugger] Improve syntax highlighting.
Adds syntax highlight support for lines beginning with # which are
formatted like a comment. This handles C preprocessor and Rust
attributes and improves source code listings somewhat.

Adds tokenizer support for Rust 'lifetime tokens. This is not needed to
support any expressions but comes up regularly in syntax highlighting.
Without this patch, any source line with a lifetime in it is
unhighlighted. With this, lifetimes are highlighted like keywords and
any other tokens on the same like are also supported.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621753
Original-Revision: 9267040a16ce79f7d15482736ffeadd37263f6ca
GitOrigin-RevId: c98a2405b419b3358bf124080be48ff7a67604c4
Change-Id: I2b531249a311bec18aa2b5bd5a6d1794f6b8fcb5