[roll] Roll fuchsia [timezone] Removes fuchsia.deprecatedtimezone.*
All out of tree uses of this FIDL API have been either removed, or are
dead code now.

Given that realization, it should be safe to remove the FIDL API and all
its in-tree uses.  This change is incompatible with branches of Fuchsia
that have been cut before the date this has been committed, since petals
will be unaware of the change.

This is the summary of changes:

- the `timezone` program is removed.  All its users are now gone.
- the code that communicates with `timezone` is removed from setui
- all code referencing `fuchsia.deprecatedtimezone.*` is removed.
- all tests referencing `fuchsia.deprecatedtimezone.*` is removed.
- the FIDL library `fuchsia.deprecatedtimezone` is removed from
  the SDK
- all transitive references to the removed library are removed from
  product definitions.
- references to timezone deprecation are removed from documentation.

Original-Fixed: 42394,42389,8859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/503552
Original-Revision: b4a2eb6bc11403b763d7dfe72955894ae666e165
GitOrigin-RevId: 286f1fc4b4999eba9b3284ce7c08725a663c366f
Change-Id: Ibd0508fa81f3446d3e7d8975de1a5057c0a2c645