[roll] Roll fuchsia [realm_builder][rust] ensure library functions are Send
The futures constructed by the Rust client library must have the Send
trait in order to be used with a multi-threaded executor. Fix an issue
where one function was not Send, and add a test that ensures all the
functions work on a multi-threaded executor.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/622530
Original-Revision: 692c27105343c0a733d9953904f085ee5de58639
GitOrigin-RevId: 6ca80b8f5a77b479b8ed1acd14b378dd4280fac8
Change-Id: Ibb99e3345850f7eac32ee5b4d4e881de393e495b