[roll] Roll fuchsia [sdk][flutter] Fix roll failure.
A new lint that is rolling in notes that forcing the description's
value in a non-nullable field shouldn't be necessary. See
https://ci.chromium.org/ui/p/turquoise/builders/global.try/core.arm64-asan/b8826565112806557921/overview

We fix the issue by defaulting to an empty string when there is no description.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/624904
Original-Revision: 4b3d8a1ff54781cd1b74f0d5d30aec0e26874c1b
GitOrigin-RevId: 13424602e66a2eb263ed31ebf3ccc8fd687a2d5b
Change-Id: Iefc972abb12a7f957ce4a840c8a45fac7cd9bf83