[roll] Roll fuchsia [display] Use configuration stamps in ApplyConfig and Vsync events.
For each display configuration applied to display Controller
service, it will be associated with a unique tag (config stamp).

Display controller clients could call GetLatestAppliedConfigStamp()
to get the config stamp for previous ApplyConfig().

On each hardware Vsync signal, the display controller will send a
OnVsync2() event to its active client. The event contains a config
stamp means that all display layers has fully applied the configuration
with that stamp.

If there are pending images in a configuration, the configuration
stamp in Vsync will be only updated after the pending images are
resolved, or the layer is removed in a subsequent config, or the
layer gets a new config that overrides this configuration.

TEST=Added new test cases in display integration-tests.
Original-Bug: 72588

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620345
Original-Revision: 9f85bee7eff8882b67d3f89002ba531e6f08d093
GitOrigin-RevId: 290dac9f95d70933ac8dcf3d1fdc8e9acb4f4144
Change-Id: Ia0dfa1d9f2530fee2c17cfda9d68249ce9e39bcf