[roll] Roll fuchsia [component_manager] Add a debug configuration for nested component manager
When a nested component manager is started by RealmBuilder, it is
usually using a configuration + manifest that does not enable debug
mode.

This means that tests cannot subscribe to events before starting the
component topology.

In the transition from OpaqueTest, we will need a config file and
manifest for a component manager that uses RealmBuilder and is in
debug mode.

Original-Bug: 71873
Test: N/A
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/624223
Original-Revision: 5fa20c9eaa8a9cf89b66b3aabf70be23d5f37c58
GitOrigin-RevId: 731a080b5e0c1849ceda23292d727cb8202baee9
Change-Id: I60cc7c71c9f6b23d4f81f25c748a890eed955e5c