[roll] Roll fuchsia [display] Introduce SetDisplayPower.
Hardware-specific drivers can implement this as a no-op for now, and
this particular API will evolve as we discover new requirements.

The core display driver will also be updated with a new API, but it is
not yet clear whether to modify fuchsia.hardware.display.Controller or
to introduce a secondary protocol.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573565
Original-Revision: b320e470a44965b5bb745cc787b8a213153b2cc9
GitOrigin-RevId: 0962dcc7a14563d05d4b626d14102fc412e2b847
Change-Id: I5aa1c7a1cb276950e2ed20b56fc760c48d9545c7