[roll] Roll fuchsia [component][fuzzing] Add fuzz-manager
This CL adds an implementation of the fuchsia.fuzzer.Manager protocol.
This service will be available on startup when included in a build. It
accepts requests from remote-control and starts fuzz tests using the
test-manager service.

Additionally, the fuzz-manager bridges the diagnostic streams such as
stdout, stderr and syslog from the test-manager to clients. It buffers
this data to allow clients to disconnect and reconnect as part of
long-running fuzzing sessions.

Original-Bug: 24740
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/558391
Original-Revision: 5a9e44abb0cfd9592c5e9ee62bea26ec1deacc71
GitOrigin-RevId: 31532add8c428b7835c5872571d6c66f1f3dd557
Change-Id: Ic0e28970aa694bc25811d56fb3307e9735665b36