[roll] Roll fuchsia [bt][gap] Re-pair when peer is missing link key
When we believe we are bonded to a peer and initiate authentication, the
controller will ask for our link key and then report Authenticaion
Complete with "PIN or Key Missing" if the peer does not have a link key
for us.  In this case, we should act as if we don't have a link key, and
try to re-pair, but also confirm a new pairing with the local delegate.

Original-Bug: b/210016847
Test: added unit test; fx test bt-host-gap-tests
Test: forget key on peer side, attempt to reconnect from Fuchsia, it
      should succeed.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/625662
Original-Revision: f99085cbdaa846890859460d92093639a62ea11a
GitOrigin-RevId: ad3c377ebf0f39b8e83b6f412406276c321eb599
Change-Id: Ib3b6e72bd50a8b54bedf29414a51e206abd2f57a