[roll] Roll fuchsia [component][fuzzing] Add fuzz-registry
This CL adds an implementation of fuchsia.fuzzer.Registry. The
fuzz-test-runner adds one end of a fuchsia.fuzzer.ControllerProvider
channel to the registry when starting a fuzz test. The fuzz manager
can then forward requests for fuchsia.fuzzer.Controller to this
service.

The implementation of this protocol is in a separate component to avoid
creating a circular dependency between the fuzz-manager, the
test-manager, and the fuzz-test-runner.

Original-Bug: 24740
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/580183
Original-Revision: 35ff94b814a4f73774a70dea095ac86f7a7e2a68
GitOrigin-RevId: 05c11e8b3224cde587a2a5a661a2374dc547cf4d
Change-Id: Ia105db72642b0b22c462e0004b17e33e7bef6933