[roll] Roll fuchsia [fdf][compat] Fix the DFv1 loader settings
The current call to service::Clone does not work on the loader
service. This means that the function currently leaves the processes
loader in a bad state. This was not discovered by existing tests because
there were no compat shim tests that were colocated
(nothing was using the loader after loading the initial driver).

This fixes the issue by calling Loader::Clone directly.
It is also verified in a later commit in the 'compat-shim-test'
by launching two compat drivers in the same driver host.

Original-Bug: 90629

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/624762
Original-Revision: 85257f9ccbf465c5c1fef6521c39575f29679fab
GitOrigin-RevId: 820f8cc7b959947d36c5bbee2d21f17c68f197bd
Change-Id: Ie480ade49c91d8c764c5d1518fd5ec0edb1cccdf