[roll] Roll infra/recipes [git] Make get_changed_files return empty list if no files are changed
Previously it would return a single-element list containing an empty
string. This was rare but could happen if the function was called with
`ignore_submodules=True` on a commit that only touched a submodule,
leading to failures such as ci.chromium.org/b/8826578963573649841 where
a build tries to treat an empty string as a file path.

Also emit a JSON log instead of a newline-separated log to make it
easier to distinguish between an empty list and a list that only
contains empty strings.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/626341
Original-Revision: 4c8dd8e2fb88824e4f87e61d9c1bef47245eb12a
GitOrigin-RevId: 8c17fda4e9174da2267efb59f7e57e5b0588394c
Change-Id: Ie0ecb10427969c40e2c350d76f2252704a0c445b