[roll] Roll fuchsia [testrunner] Add NsJail wrapper
Adds a thin wrapper around NsJail that will facilitate the construction
of command line invocations. Also adds a flag to store a passed in
NsJail binary path. This CL does not actually wire up the usage of said
flag or wrapper, it just puts the pieces in place for a future CL to
make use of them. This was done to enable CQ testing of NsJail changes.

Original-Bug: b/211504655
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/624683
Original-Revision: f53d8853104207515940e90a167390f7ee02f56a
GitOrigin-RevId: 75ae42bf9f37fcf81a3326762eb9cdc914d760b9
Change-Id: I69a5eb2da727934755be2970c5353013d0292774