[roll] Roll fuchsia [blobfs] queue incoming requests until blobfs is started
This changes the component runner to queue incoming requests to the root
of the filesystem and the svc directory until the startup protocol is
called. It does this by creating dangling channel pairs which it puts in
it's outgoing directory at /root and /svc, which it only starts serving
once the Configure callback is called by the startup service. The
diagnostics directory isn't expected to adhere to this behavior, and so
it doesn't get this treatment.

This also adds a test which verifies the request queuing behavior to
both the blobfs integration tests and the blobfs-component tests.

Test: fx test blobfs-component-runner-test blobfs-component-test
Original-Bug: 82896
Original-Bug: 39588
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/625747
Original-Revision: 936b321fec6cf85dfa3afef5a4f84c4a4bc96c4d
GitOrigin-RevId: 755d76d8f7596707e6c5f176c470c96ec59a6df4
Change-Id: I813b972e06a8e1ec366486c3897227abefe9fee8