[roll] Roll fuchsia [netdevice] Introduce virtio netdevice driver
Note that the new driver is only part of the build but not introduced to
any boards because netsvc and device-name-provider get in the way of
device discovery. We can only add the new driver to images once that's
addressed.

Original-Fixed: 89489
Multiply: virtio-netdevice-test: 30
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610875
Original-Revision: e19748f1de93a6eb40a56e367bf6777766070fad
GitOrigin-RevId: a3e2f5d973eb9ee4400a2bca53ac39ad8116047b
Change-Id: I07f56e07bb230b7fe7fc01d7d062f3c22729d962