[roll] Roll fuchsia [ffx] Add template rendering for ffx emu flags
This CL adds the template file for ffx emu command line arguments, and
the code for parsing and utilizing the template to build the command
line.

We use Handlebars to provide variable substitution and conditional
inclusion of certain flags in the template. Two new handlebars helpers
are defined, one for verifying the value of an enumeration, and the
other to call the abbreviate() function on a DataUnits member.

Original-Bug: fxbug.dev/90481

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621241
Original-Revision: 1039fc7c59bb5cd464478cd6f0311465d75e899a
GitOrigin-RevId: 8089d885b3326f581492071d464e317317592ab2
Change-Id: Ia461cc1df458a5c5a5eee6aec248cd1e90b01c86