[roll] Roll fuchsia [scrutiny] Fix directionality of routes
The documentation in the Routes collection states that the source of a
route should be the component providing the capability. However, the
collector constructs routes with the source being the component that
uses the capability, effectively reversing the directionality. The
capability source was also confusingly called "target".

This change fixes the name to make it clear it is the source of the
capability, and swaps the source and destination in the Route
assignment. It also introduces two tests to observe the directionality
of routes, which was not previously covered.

Test: fx test scrutiny_plugins_lib_test
Original-Fixed: 90927
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/626430
Original-Revision: 2d3281f417714c7c650da885a89f8be112c08d58
GitOrigin-RevId: c948b5c0b34fcb11e9f6ebfddaf913a98e1d9206
Change-Id: If67f4463f0d59c196a0290cc7a270574db973998