[roll] Roll fuchsia [fidl_driver] Arenas for driver transport
This CL wires through the driver arena parameter so that arenas can be
used in requests and responses. The API used here is not meant to be
final, but merely to make a functional interface.

A disabled end-to-end test is included in this CL. The test typically
passes on non-ASAN builds, but is flaky and involves a use-after-free
in driver runtime teardown. This use-after-free is not new to this CL
and the CL is being sent for review despite the use-after-free to
prevent code rot and to make it possible for other changes to continue
to build on top of it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621621
Original-Revision: f824a24db88a82798a449893219e0307e839ee89
GitOrigin-RevId: ef8c3c9ad8983c3650802128dd56ef8829371043
Change-Id: Iaf2419c37e321a2dc43573c3e2b830a3a8a8143a