[roll] Roll fuchsia [zxio] Add a zxio_reopen operation with flags
This adds a zxio_reopen wrapper for the io1 Clone or io2 Reopen
operation and defines an option to request a description of the
connection from the server. zxio_clone internally calls zxio_reopen with
the default options.

Currently, reopen is implemented with a fuchsia.io.Node/Clone call. In
the future we'll want to switch this over to the io2 Reopen call.

This also adds tests for both the zxio_clone and the new
zxio_reopen operations with the remote backend.

Original-Bug: 43267
Original-Bug: 77623

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627028
Original-Revision: c3ef186946280b9bd3b1a43f9d105bd18db42cc6
GitOrigin-RevId: 4f4da9b266ed168e8d4c8063e1e82baf3a03a026
Change-Id: If52229d661039672db458b85b901f3e7b6b3d176