[roll] Roll fuchsia [ffx] Additional flags in the start-up template
This CL adds more flags to the start-up template, based on
device_launcher code. Specifically, we add flags for hidpi scaling, gpu,
window size, audio, image files, and serial monitoring, then added a few
predicates for some that were previously hard-coded but shouldn't be.

Based on a dry-run comparison this takes us to parity with the default
setting in fvdl, with these exceptions:
 - grpc port flags
 - fvm image flags
 - networking flags

For the acceleration flag, this checks that KVM is accessible and that the
current user can write to it in order to resolve the "auto" value.

It also updates the sequence of events so that template resolution
happens after file staging and other tasks, but just before the command
is launched. This makes sure the flags contain the most up-to-date
values.

Original-Bug: fxbug.dev/90481
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/626985
Original-Revision: c17596f7e63a88c62dd1093f6662d0132da8363d
GitOrigin-RevId: 2bfcd56803cae199bc346edd468a25ee0c5a56df
Change-Id: I06982d7e458fe9d9010c2819c8eb8638a34db4e1