[roll] Roll fuchsia [terminal] Add 1.25 scale factor bucket
This adds a 1.25 scale factor bucket and moves the
code that is used to determine the scale factor into
the terminal library so it can be used by both virtcon
and the terminal app.

The 1.25 scale factor is useful for 1080p laptop displays
and the workstation product will be updated to take
advantage of it in a follow up change.

The terminal app will be using a hardcoded DPI value of
160 until Flatland API provides what is needed to
dynamically determine DPI.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/626662
Original-Revision: 4eca6a23b62e19ddc9c27a976086878587f458f9
GitOrigin-RevId: 8a13d2ed321aff3dbbe9269e06c1289584b2d7e0
Change-Id: I53abb98b927b9979c0ef9a31e485f75490d3ae88