[roll] Roll fuchsia [fdio] Use zxio to interact with remote nodes in local namespace
The fdio library maintains a local namespace of paths that terminate in
remote objects that communicate with external fuchsia.io servers. This
updates the local namespace logic to use the zxio library for the server
interactions to better encapsulate the FIDL logic. This does increase
the size of a LocalVnode instance to store an arbitrary zxio object
instead of only a channel handle.

This also cleans up the type management in the local connection tracking
a bit to use constructors, destructors, and RefPtr for memory management
instead of raw pointer manipuation.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/626801
Original-Revision: 38873648019dee5ce350a292342ec908ac548f70
GitOrigin-RevId: 843cea608f28d814ee4a8e16da4024dd69a2c471
Change-Id: I072b24ad5f21502b2200fc0fbf3a7b170685608f