[roll] Roll fuchsia [scrutiny][resolvers] Scrutiny uses PartialAbsoluteMoniker for component comparison
Replace `AbsoluteMoniker` with `PartialAbsoluteMoniker` when Scrutiny
compares an instance moniker with a request moniker while walking
through the component graph. Additionally, instead of translating the
monikers into Strings for comparison, the monikers are compared
directly. Push the `PartialAbsoluteMoniker` of the visited instance in
Scrutiny's moniker list, since it does not include instances already.

This change has a dependency that all components in the component graph
have unique `PartialAbsoluteMoniker` values so that accuracy is not lost
when using it for comparisons. The reason for this change is to remove
AbsoluteMonikers from the public API space.

Original-Bug: 87407
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627000
Original-Revision: c897bf4d28e41be34b98d64d39279414f25fb46d
GitOrigin-RevId: 5cccff46546c5bf876364a3bd68323addc564e51
Change-Id: Iabb73d92363851801dd31cdb6ff39450fa7ceaf8