[roll] Roll fuchsia [fidlc] Output transport in JSON IR for client end / server end
In order for fidlgen generators to know if a client end or server end is
associate with drivers or zircon channels, it is necessary to know the
transport. This CL outputs the transport alongside the client end and
server end in the JSON IR in order to accomplish this.

Original-Bug: 56727

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627201
Original-Revision: 001c6358cd1c9fee2a2aae5b832f99d0a1d18c81
GitOrigin-RevId: 2ecace2c8c3a7880e1b6484cb2ba6c50c9f55dfc
Change-Id: I32e8a801fb0e30fa074bb8f2c510735f73c7dd1a