[roll] Roll infra/recipes [checkout] Disallow using start_dir as checkout root
Recipes tend to use start_dir as a dumping ground for various temporary
files and CIPD packages, which pollutes any checkouts that are rooted in
start_dir.

Rooting a checkout in start_dir also lets a recipe implicitly run later
commands in the checkout directory without using `api.context(cwd=...)`,
making the recipe implicitly dependent on start_dir being the checkout
directory. It's better to force each recipe to be explicit about which
commands need to run in the checkout directory.

And since it can be annoying to choose which directory to use for a
checkout, add logic to construct a default checkout directory name based
on the remote URL.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/627641
Original-Revision: 624c51083614f565d8daf82c1d98f118cb02e60b
GitOrigin-RevId: 077da9fb83d2d9584869de6afdedaa05468f5053
Change-Id: Ic27708765e7ccb1d1ceacffb6328dd9b9347edf8