[roll] Roll infra/recipes [git] Disallow using start_dir as checkout root
(Applies the same fix to the `git` module as fxrev.dev/627641 applied to
the `checkout` module.)

Recipes tend to use start_dir as a dumping ground for various temporary
files and CIPD packages, which pollutes any checkouts that are rooted in
start_dir.

Rooting a checkout in start_dir also lets a recipe implicitly run later
commands in the checkout directory without using `api.context(cwd=...)`,
making the recipe implicitly dependent on start_dir being the checkout
directory. It's better to force each recipe to be explicit about which
commands need to run in the checkout directory.

And since it can be annoying to choose which directory to use for a
checkout, add logic to construct a default checkout directory name based
on the remote URL.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/627642
Original-Revision: d12c20c50dad0c8b485a8d57dfa92acb64964fd9
GitOrigin-RevId: 8599a2e7da1473aa8c30b479f3858539e124afcf
Change-Id: I8c6157a88d40e646f8fe50f2da29fbd9c88f6d40