[roll] Roll fuchsia [Recovery] Add new Proxy View navigation controller
ProxyView allows the application to switch views enabling
completely different screens to be shown.

ProxyView pushes new views to a stack and pops them off when
requested by a Carnelian Message.

This is the first of several connected CLs for the on-screen
keyboard which uses ProxyView to show and hide the keyboard.

A change has been made to the carnelian sources because I could
find no way to create a mock ViewAssistantContext.

Original-Bug: b/202857033 b/203107208
Test: New tests added
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620081
Original-Revision: 521adcfd579a7b6b4a011848478576378c0d8478
GitOrigin-RevId: 0627dcc4be53b652cc1455d82d0ac2346a39f1d1
Change-Id: Ia852274f48c92dccb7cf47b871488d028d6cb425