[roll] Roll infra/recipes [build_input_resolver] Delete
build_input_resolver was originally written to make sure that specs were
fetched from the same base commit as the main checkout. This is no
longer necessary now that specs are just regular properties.

Even if that wasn't the case, recipe_bootstrap resolves a Gitiles commit
for every build anyway, so removing build_input_resolver is a no-op and
won't break the few places that assume the build has an input commit;
based on the Buildbucket BigQuery table, build_input_resolver hasn't
done anything on any build in the last 6 months, with the exception of a
small number of misconfigured tryjobs that appear to have been launched
manually in November 2021:
https://plx.corp.google.com/scripts2/script_61._d73c29_0000_22bc_88d0_582429a7900c

build_input_resolver was also a bit of an antipattern because it
modified the global `api.buildbucket.build`, which the buildbucket
recipe module explicitly says NOT to do:
https://chromium.googlesource.com/infra/luci/recipes-py/+/95dd09206fc411ea321f57ca3905f9785b7ef44d/recipe_modules/buildbucket/api.py#138

It seems that build_input_resolver ended up being used in so many
recipes because it simplified recipes that had separate checkout
handling for try and CI builds, by letting them assume the build has an
input commit (e.g. see fxrev.dev/405457). But all such recipes now use
the `checkout` or `git` recipe module methods that have handling for a
missing gitiles commit.

Original-Bug: 77879
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/627801
Original-Revision: b12144343fcdc99590be3c8cdb1fde895a9b48cb
GitOrigin-RevId: f432769a3c6c93e14d4852f5d4bbc94a83becf42
Change-Id: I08b148ba0c689ff14fb488afd06ec43596ffb435