[roll] Roll fuchsia [fidl][llcpp] Some helper methods for fidl::Reason
fidl::Reason is an enum with lots of members, some of which have subtle
meanings. Having an enum could be problematic for future extensions
because users might try to exhaustively match on the members.
This CL is an interim step to address the complexity. While one could
still reach for the underlying enum and write `switch case` if they
really need it, it is now discouraged, in comparison to a few new helper
accessors.

I've changed most of the big `switch case`s in-tree to use the accessors
to demonstrate the simplification, but will split user code changes out
when landing.

Original-Bug: 75702

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/626807
Original-Revision: 2d7a98864d1ff43430c8f718b0961776443c4416
GitOrigin-RevId: ff475849c3df2f895ba7c03ed4439b3bd7e1b52b
Change-Id: I3e41889f7fa76a3721850d6538e7f75674b48f45