[roll] Roll fuchsia [ffx] Move environment vars to the args template
This CL adds an "envs" field to the FlagData struct, so that environment
variables can be specified in the start up args template file. We remove
the previously unused env field from the runtime struct, as well as any
references to it. Then we added a few defaults to the template for
DISPLAY and QEMU_AUDIO_DRV, which fix errors or warnings we were seeing
without them, and a template helper to populate fields with environment
value.

To aid in correct DISPLAY behavior, we also output a hint to switch to
--headless if none is set (which usually indicates a text-only session).

Original-Bug: fxbug.dev/90481
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627765
Original-Revision: 09310f1ed3aa22576a496b3904a10bf3a997a7b6
GitOrigin-RevId: cb6df14d6f3c67af22285bc2ad908ef615778829
Change-Id: I4d763d21bf9d360bdf713b20872b4ed362dda0f3