[roll] Roll fuchsia [llcpp] Split out wire messaging declarations
There is a strange ordering of dependencies, where wire_messaging.h is
both the top-level entry point header for messaging and a dependency
of other headers. This makes it harder to avoid dependency cycles with
includes. This CL splits out the forward declarations into a separate
header to reduce some of the complication.

This also fixes a bug, where a call to fidl_write_epitaph may in some
cases be made for non-zircon channel handles.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627464
Original-Revision: 4014d699db49358509f78e0fa525a99a0009bbdf
GitOrigin-RevId: 832f137ecffb08a2a9292157c9acc59f7c843397
Change-Id: I6eab99e70b65557b733bdac5f10810ee2c95dab5