[roll] Roll fuchsia [llcpp] Support driver client end inside of FIDL types
This CL makes it so driver client_end and server_end, specified in the
FIDL language are generated into fdf::ClientEnd and fdf::ServerEnd.

One of the bigger challenges is dealing with a dependency on the driver
library from the wire type library. To solve this, a flag was added to
the FIDL build rule indicating that the FIDL files use drivers.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627462
Original-Revision: 24d21e3363144c06fd75db01b3c4b8d649c5547d
GitOrigin-RevId: aef6a5e8d68e2926df629ba63bbf0cfbc6b39352
Change-Id: If9e4f4b78b3edf74509da636c78fa154d711e8c6