[roll] Roll fuchsia [Recovery] Add Keys and Keyboard layout code to recovery
The keys have lower, upper case and alt variants.

The keys are brought together in the keyboard which also
contains two text fields, one for a title an one for user input.

This is the second of several connected CLs for the on-screen
keyboard which uses ProxyView to show and hide the keyboard.

Original-Bug: b/202857033 b/203107208
Test: None
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620084
Original-Revision: 38f945099ac5736a66474b2832784b80f046bf6a
GitOrigin-RevId: 454f44e26ff2877bc5282d7aed55f751ae9f8ddf
Change-Id: I2c7a0b8ea58d67df35d463a5d16821a1aa67d092