[roll] Roll fuchsia [Recovery] Add buttons for WiFi connection
These buttons do not do anything yet but are displayed on
the bottom of the screen. They will be hooked up in the next CL.

This is the third of several connected CLs for the on-screen
keyboard which uses buttons to show the keyboard.

Original-Bug: b/202857033 b/203107208
Test: None
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620163
Original-Revision: a889ba31141f7f1c88b7875afdcf3f95114badd8
GitOrigin-RevId: a35e39d52d8012b51fe5c2c213a5ddc0323905df
Change-Id: Id7a2519abcb63e987bd32ca5c05e80196e3dd374