[roll] Roll infra/recipes [auto_roller] Pass if changed already pushed
If one push attempt times out but the push actually succeeded on the
server side, later attempts will fail with a "no new changes" error
message indicating the change is already up-to-date with the local
commit. Example failure: https://ci.chromium.org/b/8827921599682608289

That means that a previous push succeeded and the change is in the
desired state, so we should proceed normally in that case rather than
failing the roll.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/628441
Original-Revision: 964753b9f5a23d3c8877e2a53c2500c05c073a05
GitOrigin-RevId: e0cf4d77a064827b70998d178df5aee24de2ba18
Change-Id: I84dec2e2c59f50098a11012dd661c2cdfa8d5337