[roll] Roll fuchsia [media/vnext] Update to fuchsia.audiovideo.Producer
This CL updates the FIDL definition for fuchsia.audiovideo.Producer
to address issues that arise when requesting compression types other
than the default 'native' compression type.

An A/V producer (a demux, basically) can now alter the compression
on produced streams. This required some updates to the FIDL
definition.

1) ProducerStreamInfo, which describes a stream, is now a
table and describes the native format and compression as well as
available alternative compression types.

2) The ProducerStream definition has also been updated to have a
separate SetCompressionType method rather than just taking Connect
parameters that specify compression. This was done because the format
produced after a decoder is added may be different than the original
format and because, in encode/transcode cases, the new compression
may have a parameters blob that needs to be specified in the connect
call.

Test: fidl changes only
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/626556
Original-Revision: 627d5f64a7005a18feee1269814c4df93655ac46
GitOrigin-RevId: ba36f1e732606bbfbda1e69ce13b2e7e46c2c5f3
Change-Id: Iddb304c307545dffa54412d1c671bc16fc45b1ee