[roll] Roll fuchsia [ffx] Add flags for TUN/TAP networking
This CL adds the basic flags for running the emulator in Tap network
mode, by adding a tap interface on the default mac address. It makes no
effort to set up the interface, but prints an error if it's not already
configured. It will also print a warning if Tap mode is attempted on
MacOS.

Manual testing shows this is sufficient to enable a user to access the
emulator through `fx shell` on a Linux machine with a Tap interface in
place.

Original-Bug: fxbug.dev/90481
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627923
Original-Revision: 68cf5a7538ece4a92911bc85729a6cbd68a49c86
GitOrigin-RevId: a254fd0f17ad67a14a1bf7b4fae62969f12c3d23
Change-Id: I46fde90911a6e2a545857fb35e0d2a857ad2ec14