[roll] Roll fuchsia [cursor] Add cursor component
This implements a simple cursor component using carnelian and
makes the Flatland scene manager use it to display a cursor.

The cursor component produces the appropriate output for
a cursor be rendering a rive file. The component can be
updated to support different cursor types and animations
in the future.

The Flatland scene manager implement a cursor sprite by
embedding the cursor components view, and controlling its
size and location in the scene to match the location of
the pointer.

Carnelian has been updated to allow clients to disable input.
This is required to not have the cursor view steal all input
until we have a better way for the scene manager to handle
this.

MouseInjectorHandler has been updated to match the gfx mouse
handler by centering the cursor initially and supporting
cursor hiding on touch and immersive mode.

Test: input_pipeline_lib_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/613242
Original-Revision: 916c0a03dd60cdf6b4fc9eb8d60988a56a6e0de1
GitOrigin-RevId: 8aa03149bae45664c879ec5d8e427246ffaa6ded
Change-Id: I9fa6682c7caf6b7a4d3e6ee31e9db4f1aefed1e9