[roll] Roll fuchsia [fidl][drivers] Move client APIs behind `.buffer()`
This aligns the driver client closer to the regular fidl::WireClient, so
that they both use `.buffer(...)` to pass in an arena, and get back an
interface to make calls.

This is accomplished by creating driver-specific client wrappers,
fdf::WireClient and fdf::WireSharedClient. They are more-or-less copied
from their fidl:: counterparts, but with bits and pieces tweaked to be
more driver specific:

  - Takes fdf_dispatcher_t* instead of async_dispatcher_t*
  - `.buffer(...)` takes a `fdf::Arena&&` instead of any memory resource
  - `client->FooMethod()` is not supported. We don't have an immediate
  need for the managed memory flavor.
  - Some comments are shortened to reference the fidl:: counterpart
  to reduce future matainence burden

TEST: tested that DriverTransport.DISABLED_TwoWayAsync works up to
canceling the channel wait when un-disabled.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627756
Original-Revision: 1e3c90d95131c637719c46662a19d13003e4bdeb
GitOrigin-RevId: b76e1b7f5e4c67619527d8c389fae196a59fcd91
Change-Id: I35b24e9de7faa3dc9ef202757302f42a629e6b1f