[roll] Roll fuchsia [testrunner] Handle non-fatal errors within Test().
I added a TestTimeout status and a FailReason to the testrunner
TestResult so that we can set those values within the TestResult
returned by Test() and have it only return fatal errors. That way,
main.go doesn't need to check the type of error it gets from Test() to
see if it's fatal or not.

Original-Bug: 49266
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/628089
Original-Revision: 2f3661dcc83ab9959a86819b3ffb3c922786f83f
GitOrigin-RevId: a8610c1d89c447a5b55d261177e17ff6602d093d
Change-Id: Ibff629919592870ba60dcd71f2660269021c9df2