[roll] Roll experiences [login] add a hosting directory to hold account_data
The flutter runner expects that all top level directories
exist at the time that outgoing is served since it
is using a composed directory. We have to put in one
more level of indirection since account_data does not
exist when out/ is served. This allows us to add the
account_data at a later time after the user has logged in.
BUG: 89690

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/628505
Original-Revision: 1010b96329033e25cb824ac0235b0b7139f83389
GitOrigin-RevId: 7f6fe6cd4f4fabf91864afab4573d085fb957574
Change-Id: I6b4e8c13fced9270f4087486e74274e72d949288