[roll] Roll fuchsia [fidlc] Provide Spans for ErrIncludeCycle
Update call sites for ErrIncludeCycle to use Fail instead of FailNoSpan,
providing the span for the name of one of the Decls in the include
cycle. Also provide a listing of decls participating in the cycle in the
error message.

Original-Bug: 89213
Test: fx test fidl-compiler
Test: Added test cases to exercise some edge cases in BuildExampleCycle.
Test: Manually tested error message formatting.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627473
Original-Revision: aef2444612dabbf5e1ac7c31a32831fd5c511c39
GitOrigin-RevId: 34b2403953c9ecf33ecf0c23ac06937d1353ff1f
Change-Id: I28693e43295119c8567af665c887bd1b9b152cf6