[roll] Roll fuchsia [build] Add fastboot bringup image to images.json
Booting bringup images with fastboot requires concatenating the ZBI with
the vbmeta. A target that does this was added in a previous change, this
change just adds the resulting image to images.json so infrastructure
can use it.

Original-Bug: 47531
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/628093
Original-Revision: 9cc136473663e96a46b762db98c447e03eac8c8c
GitOrigin-RevId: d13943800f3fd0faa24c483629ad24a5262d60bc
Change-Id: Ife23ef993d6d0b9db29ddb9f0b7630c98e43ba18