[roll] Roll fuchsia [kernel][vm] Extend dirty requests over lookup range
When generating ZX_PAGER_VMO_DIRTY requests, try to extend the range to
include other clean pages and zero page markers that are also going to
need requests. This is an optimization to reduce the number of distinct
requests that need to be sent overall.

Before returning from LookupPagesLocked, extend the range to cover the
max lookup length, regardless of how many pages were actually looked up.
This is fine, because we are not returning any looked up pages when a
dirty request is generated. We are simply forwarding the request to the
user pager and returning with ZX_ERR_SHOULD_WAIT.

Test: core-pager-writeback-test

Original-Bug: 63989
Original-Bug: 88674
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627228
Original-Revision: ac4d327b4c1c6f9852ff2faa2b800d3c7456fdc3
GitOrigin-RevId: 35d9d671bfac3618ac3e2d37f97a7a35c4437169
Change-Id: Ibed79e57620d9d4b54e1bcd4db2ccbce06d4f38f