[roll] Roll fuchsia [kernel] Add EntropyPool::AddFromDigest
Adds logic to entropy pool to add the digest of input
bytes instead of the bytes directly.

This will be used on phys, so validated hex strings are
added.

The size of the digest is treated as an implementation detail,
so it is returned, so the caller can perform any entropy estimation.

Also made a public config for the include directory.

Test: k ut crypto-entropy-pool
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/612402
Original-Revision: 5bc45b9138a2fe55ca01b7495c1328beb063d886
GitOrigin-RevId: 09f3b84fcb8811bcf64105d13a7bbd0cf7e5c548
Change-Id: Id5eefc676ff33d8fa3b139f670eac16941ed742a