[roll] Roll fuchsia [testrunner] Mock ssh.ExitError
Instead of unwrapping an error as an ssh.ExitError to check if the test
command had a retcode that indicates a timeout, unwrap it as an
interface implementing only the subset of ssh.ExitError's methods that
we need to use. This makes it possible to create a fake error type to
test handling of timeouts without having a separate testing-specific
code path, which was previously necessary because it's not possible to
construct an ssh.ExitError in-memory due to its fields being private.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/629181
Original-Revision: 91b7cda88f980eee8cc9d6044369b0c1462136eb
GitOrigin-RevId: 3d4271152bfe96d390928f20b16ebc50c7f305e3
Change-Id: I806b3b82e230c1d38d573e3a5a49ae1611cb78a7