[roll] Roll fuchsia [tools][fuzz] Retry failed SSH connections from undercoat
In addition to the retry logic added in the previous CL that is specific
to initial instance boot, this CL adds retry behaviour at a lower level
so that it will apply to all SSH commands run against the instance.

The two levels of retry logic remain separated because `start_instance`
needs to handle the edge case where the connection succeeds but the
remote command fails.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627880
Original-Revision: cc656b298250ffafd145a644ce9c2457df42b445
GitOrigin-RevId: 6300363df22416d6dfb21d882c052a42c29d2288
Change-Id: I3cbde35c991e46ae9b0f8128299347bafa7c73d4