[roll] Roll fuchsia [debug_data] Make data processing accessible via fidl
Creates a new 'debug_data_processor' component that contains all the
VMO processing logic. This allows us to call it via fidl from a
component written in Rust. The component is not used yet.
Note that infra scripts currently rely on the location of the debug
data on the target's filesystem for pulling coverage. To support this
case, we allow the caller to pass in a directory the processor will
work in. We can freely change this detail once the scripts are
migrated.

Tests: new unit tests
Original-Bug: 77634

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/622793
Original-Revision: e2090538d0f4f947cfcc8d30e4db36a14b95206c
GitOrigin-RevId: 38c8c7cfd36af8e94c1c6c9ec6419894ca61418a
Change-Id: I0c31366ce691cb323a5e547f51b43ca6e0d21923