[roll] Roll infra/recipes [testing_requests] Add nsjail-root to testrunner command.
nsjail needs the root of the directory to mount, but we run the
testrunner command in a relative cwd to the swarming root directory
because some of the paths to test deps point to relative paths that
start with "..". Thus we need to give nsjail the root swarming working
directory so that all those files get included as well.

Original-Bug: 91198
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/629533
Original-Revision: c268846ef14ed67ca50acb2c2c676cc80f4bf55c
GitOrigin-RevId: 3f9b4c001c2279e4e17a33fa072dfe272f6ef319
Change-Id: I425045033976222d7e0dd15221d8909e17ba5037