[roll] Roll fuchsia [rust][cf] Remove unused Capability::event method
This CL removes an unused method to unblock the Rust toolchain roll,
since it triggers a dead_code warning in the new toolchain. Although the
method is marked `pub`, the mod is private to the crate and the method
is never used.

Original-Bug: 91243
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/630101
Original-Revision: 3ce76f148c568b3a0701e1e440be90f3dbcad130
GitOrigin-RevId: 86039e988071b718f60a1a67e899efa6cf05f36e
Change-Id: I6f426d2504c2d39248e356433d383d4b4d14ff39