[roll] Roll fuchsia [bt][l2cap] Explicitly memoize FCS computation
Add code to explicitly memoize the "per-octet" subroutine of ComputeFcs
by filling a look-up table at compile time using values from
ComputeRemainderForOctet.

Update some documentation to remove references to logical LSb in the C++
computation model since that is the opposite significance to the modulo
math. Instead just refer to "left" or "right" and only mention LSb and
MSb in reference to the Bluetooth spec.

Fix bug in the way that unused_remainder is computed (the result is
correct but if the previous expression were ported to uint32_t for a
32-bit CRC then it'd produce wrong results, which is counter to the
point of documenting the code).

Original-Bug: 91277
Test: Refactor only. Ran bt-host tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/629721
Original-Revision: 28a29b1b01742c2d5bbff4b221eef2de740c41c2
GitOrigin-RevId: b93d2532b41e07d4b9b0b7c7aa26de26c6419864
Change-Id: I978531fd89951a778f9ae93279a306cff157c784