[roll] Roll fuchsia [llcpp] Remove Fuchsia-specific code from transport.h
The main thing requiring __Fuchsia__ in transport.h previously was a
reference to the async_dispatcher_t. However, this type is only needed
in transport.h to forward to other things and can be redeclared.
In the future, it may be desirable to migrate to something less
fuchsia-dependent than an async_dispatcher_t but given that it will look
identical to an async_dispatcher_t, it is easy to stick with the
existing type for now.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/630964
Original-Revision: d73d5ce742ffd426b847e093f026f03ed05bc050
GitOrigin-RevId: a9cd685b4a97a7a0f9e5f80e04f909dbcea93bdd
Change-Id: I5eae0d2a1cc657c3dfb0cf0ae2b19abe95498286