[roll] Roll fuchsia [pkg-resolver] PackageFetcher type alias -> wrapper struct
This change is entirely a no-op in hopes of making the package fetch
queue a bit easier to follow. The idea is if PackageFetcher is a wrapper
type it'll be easier to control-click your way to find the inner queue
implementation.

Because control-clicking already would take you to the type alias, which
was right next to the construction function, it's only a minor
improvement, but it still seems worthwhile.

A pleasant side-effect is the doc comments get a bit nicer to read,
since they're a lot less abstract (since the queue's processing
behavior is known rather than being arbitrary)

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/628587
Original-Revision: 44cb59ceb73edee6ad19dc2ef93c4767ef9fe0e1
GitOrigin-RevId: 884cd71027be59598b5a2538786345503cf52520
Change-Id: I81ec10ca40d94bfd1762730d66c3e4a31d58f31b