[roll] Roll fuchsia [build][fdf] Add build flag to enable dfv2
An build flag was added which allows us to add a GN argument which
enables DFv2. Eventually this will be turned on on a per board basis,
but the GN arg allows us to enable it for any board while we are
developing support for it.

It's worth noting that no board will successful boot at present if DFv2
is enabled.

Original-Bug: 90539
Test: Add `enable_dfv2 = true` to //products/bringup.gni && fx qemu
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/631265
Original-Revision: 726c19adcea5be3aff523c6b1d26c5d11935492d
GitOrigin-RevId: abba4c5a416c1a2824d7a82b37e49e47dce749c1
Change-Id: Ia091ddb016d129cb9f60ae495857e3f8416c4205