[roll] Roll fuchsia [audio][tools] SigGen specify PTS for first packet
The CLI utility signal_generator can now specify the PTS value
for the first packet it submits to the AudioRenderer. The cmdline
flag '--packet-pts' indicates that timestamps should be applied
to each packet; it is now extended to accept a value that will be
used as the PTS for the first packet submitted. This flexibility
is useful when interactively exploring cases where the packet
stream starts after (or before) the renderer starts playback.

fx shell signal_generator --rate=48000    --pts-unit=48000/1 \
                          --media-start=0 --packet-pts=48000

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627877
Original-Revision: 2e43903dc1e8fdc040ab39e51e6981fb49146976
GitOrigin-RevId: 85f2d5d0038a8db3f5aaccc423af725647fc769c
Change-Id: Id15613d56f634cc42d990344d4ab9d5274f7365a