[roll] Roll fuchsia [starnix] Set process debug address
This change sets the process' debug address to the one present in the
DT_DEBUG entry, when such an address is set by the linker.

If the process' debug address is set to `ZX_PROP_PROCESS_DEBUG_ADDR`
starnix will issue a software interrupt when the property is set to
inform the debugger.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620722
Original-Revision: 04e405a275ef05b5c5ad11af0fb88159590fddf8
GitOrigin-RevId: e3a8539e4b7a699a294d64349d8b46b465455553
Change-Id: I420c884ddd4000cc287aaf345f0aea9ca20e4c3b