[roll] Roll fuchsia [pwauth] Return more specific error when derive_key fails
When an account is created with an empty password, then it is bound to
the "NullKey" account metadata variant.  If you attempt to unlock that
account with a non-empty password, `derive_key` fails, since otherwise
there's nowhere else for us to check the password, but the FIDL error we
return for that failure was Unknown, which is confusing.  It should
instead be AuthenticationFailed.

Original-Fixed: 91125
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/631083
Original-Revision: 9a3bddf081b05b3173b2819ca0a1cae0bfe45903
GitOrigin-RevId: 43e2ff0a88bee684b05183dee7ea921603ae50e1
Change-Id: I3f2a5d0fb7bae6ec3c8d815690ac9d0257b81c79