[roll] Roll fuchsia [pretty] Add support for pretty::ParseSizeBytes
pretty::ParseSizeBytes does the reverse, to some extent, of
format_string. ParseSizeBytes will take a formatted string
and will return the number of bytes the string tries to encode.

e.g.
64k = 64 * 1024
64.25k = 64 * 1024 + 256

As commented in the source code, not to remove support for larger
units, and to be able to convert strings formatted from the library
manual software division and multiplication was added.

This is necessary sin a perfectly valid and representable size
10.4492187500t (and above) could cause the simplest approach to
overflow. To work around this, we add on element of the mantisa
at a time, and keep track of the accumulated carries from each
scaled element. Essentially this is reproduceable with any long
mantisa.

Original-Bug: 76151
Test: fx test pretty-test-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/614122
Original-Revision: 1b156f8a560fafb5b75706ad5d3725040b5bf662
GitOrigin-RevId: c823483ae75c3145ed860f50d27ada6b67f0f75a
Change-Id: If82e172577a6d33e9ea9e9fea36072ea70d2e83e