[roll] Roll fuchsia [ftl] Set up some basic unit testing for the ftl.
Sets up the ability to reach inside the Volume and underlying ram
storage to be able to inspect and manipulate the current state in order
to introduce desired failure modes.

Tests added here include a partial page write example as it is a
previously experienced issue, as well as a couple of behaviours that I
was fairly confident in, relying upon, but not really certain about.

Original-Bug: 38433

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/628091
Original-Revision: 1bba90494430d3d77353f33d8928f5d56c11d10a
GitOrigin-RevId: 347f4ea553f7b0fb7fbf0ff5c9575c3a9aa7e9da
Change-Id: I343e8bd2e65d58950edaa254205fa2d11354853d