[roll] Roll infra/recipes [git] Set default git timeout to 10m and lower of some operations
This provide a ceiling for what is acceptable. We can adjust as needed
and to one-off for surprising operations. In practice most operations
should take a handful of seconds but starting with a high ceiling as a
start.

Many operations have a default of 1 or 2 minutes, depending if they do
remote or non-trivial operations or strictly local. In practice GoB can
be flaky and we want to allow time for git-retry to do its thing.

Use 20 minutes for fetch, since it can be fairly slow (>10min) on cold fetch. It
happened in one of the try job.

Original-Bug: 91396
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/630921
Original-Revision: dc0993fe2fcf639c8b28c55d78e69a29d749b118
GitOrigin-RevId: a7e014347f8cb8b328a23e2b0a23d71e9bafea0e
Change-Id: I6bde176b34be4808de24e42c84c25e24c50e59fc