[roll] Roll fuchsia [kernel][x86] move calls to mp_set_curr_cpu_online later
Currently, x86_init_percpu() calls mp_set_curr_cpu_online(true) at the
end of the routine, which marks in the higher level percpu structure
that the cpu is online. Trouble with this is its prior to the mp
structure being constructed via global constructors so it's possible
for the bool to be overwritten. This doesn't happen currently, but in
another upcoming CL the mp struct will overwrite and cause at least one
assert to trip.

It's not really correct for that routine to set the online status
anyway, since it technically has nothing to do with the concept of cpu
'online' so move the call out to a more appropriate spot.

Since there are a few paths (boot cpu start, secondary start, resuming
from hotplug) that have been relying on the implicit call, also had to
make explicit calls.

ARM64 is unaffected by this since the cpu_online call happens inside the
per-cpu GIC setup. Thats a bit more correct, since after setting up the
GIC is the first time IPIs can actually be delivered to the cpu which is
the primary function the online bool serves.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/630281
Original-Revision: 027f34a4d36930f6964fc555dbe79bf42296ac51
GitOrigin-RevId: 7c84922b9445ac2234251eb4c297c1a377f74351
Change-Id: I7e97ffb5229ebc0cc1f4215cee0ab745c4a9c711