[roll] Roll fuchsia [testrunner] Don't write test data into the source tree
Previously the testrunner main_test would inadvertently write out files
relative to main_test.go when run with `go test` because the fake test
output dir was set to an empty string.

And enforce that the outdir passed to `CreateTestOutputs` is not an
empty string, so it's safe for the TestOutputs to assume that OutDir is
set.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/631701
Original-Revision: f2d770fc45652460a6fc59647c0f588f36ca2bd4
GitOrigin-RevId: 6369c42fde09e56394b8cab33969806931aef3d4
Change-Id: I6f8d55caa78c828906d8e8534b50623407689d3c