[roll] Roll fuchsia [system-update-checker] Fallback update channel to ""
Previously, the system-update-checker upon startup would try to discover
the update channel by:

1. Check if vbmeta contains an `ota_configs` channel.
2. Otherwise, use the rewrite engine to rewrite
   fuchsia-pkg://fuchsia.com/update/0, and use the rewritten URL's
   hostname as the update channel name.

The downside of this approach is that it's not stable. If the rewrite
rules for `fuchsia.com` are changed and the device is rewritten, the
device will think it's now on a different channel.

This patch changes the behavior to use "" as the current and target
channel if one is not discovered in vbmeta. This better reflects the
fact that we can't reliably know the origin of the current build.

Original-Fixed: 91158
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627872
Original-Revision: 8c5d7586323a215627e0c22a50515f50c5f4ec21
GitOrigin-RevId: f135c781c7677b31c5a6c4490c753df6c69105b7
Change-Id: Ifba1185a665404239ff974b4897fc8cc3f4f31ec