[roll] Roll fuchsia [connectivity] Work around MemberPointerTraits issues for C++20
`MemberPointerTraits` relies on the given classes being
default-constructible, and many of the classes used with it have deleted
default constructors. This turns out to be an error in C++20 but not
C++17, so we just give those classes default constructors.

Test: fx build with experimental_cxx_version=20
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/461726
Original-Revision: a640959ab6964df6cb5edded64af4a827bfd709a
GitOrigin-RevId: 479dfb8537fc6ad484c5f4a7b9cbb5bec7528134
Change-Id: I315b736a1779b4078bd341c235775f67b5e5cbfc