[roll] Roll fuchsia [fidl][cpp][natural] Reorganize struct template
Two changes here:

The name of the type that holds the actual data is renamed from Storage
to Storage_ and made private. This will help prevent conflicts with
nested types.

Most of the private parts of the class have been moved to after the
public parts. This is to follow C++ coding guidelines. The only private
declaration that's above the public ones is a forward declaration for
the Storage_ type since it's referred to in the public API.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/631446
Original-Revision: 0256dbdb39eca6dea37799c3199e437694e4fa86
GitOrigin-RevId: 1592b3fe78996c8e43d4578c8e9ac836d7074b68
Change-Id: Idafa6d61771d9c7c0f55a4109172486e860d2d40