[roll] Roll fuchsia [fxfs] Begin support to launch BlobFS within FxFS
BlockServer needs to be updated to handle requests from both the
Block and fuchsia.io.Node fidl protocol. In particular, the
clone request needs to be addressed which is sent when formatting
BlobFS. This commit handles the clone request which creates a
new BlockServer that shares the same FxFilesystem. Stubs are
placed for the other requests.

Original-Bug: 89873
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/629684
Original-Revision: 11057550db2955602df6fea8fde2835329514527
GitOrigin-RevId: 570b105c8338d1d46eedddea25f9073b1fe55da1
Change-Id: Ia7612fbe1eddbc535e6668fe7c5a838e4a447539