[roll] Roll fuchsia [Bootfs] Add a Rust system test.
This adds a system test that briefly checks that the expected directories
exist, and that files in those directories have the correct permissions.
This is based off of (and will replace) the C++ integration test paired
with bootsvc.

Please note that this test adds the '/boot' directory capability to
system tests, which are allowed to be non-hermetic. If we don't want to
extend this capability to all system tests we'll want to add a static
test child instead.

Original-Bug: 44784

Testing:
This is a test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/626182
Original-Revision: 0a963752dbc93a0c4c3adab040a06008018b237b
GitOrigin-RevId: 5ee11dae33a723c33b02cf41ace903dbb0159e1b
Change-Id: I5dd958a0dd052c245f69529808957dc027129289