[roll] Roll fuchsia [lowpan][openthread_plat]: Remove config requirement for ResetConnection
In commit `7242f6d6949a9c`, the method `ResetConnection()` was added
to `SpinelFidlInterface`, but the addition was gated on the presence
of the config macro `OPENTHREAD_SPINEL_CONFIG_RESET_CONNECTION`.

It's unclear what the original intent of gating the existence of that
method was, and it now looks to be an oversight. Since it is causing
problems with updating to more recent versions of OpenThread, this
commit removes the macros so that this method is always defined.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632388
Original-Revision: bfca4a19a31e3473a6388feef1fd66a90c41ff0d
GitOrigin-RevId: 4d583ac9d7bb50afc15ad2312ff29b9104d44b98
Change-Id: I25e121bbc75b56adebf324a829a0a0e7b65c449e