[roll] Roll infra/recipes [rbe] Force `instance` property to be set
We can't rely on the `instance` value provided by the checked in config
file in fuchsia.git. I was incorrectly assuming that RBE would fail on
RBE_instance="" however, they ignore the value and fall back to the
config file. This means we need to assert that it is always explicitly
set.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/632401
Original-Revision: 48a188b9978d2a525ae364c62299d61e05bbc3a2
GitOrigin-RevId: 385b701497503adb7ec5d75182dfd8ee9018061d
Change-Id: I72480efb7559b2eb9ef35f1036f05f8272738c40