[roll] Roll fuchsia [ffx] Add user mode networking to ffx emu
The ffx emu plugin has been operating without user-mode networking for a
while, as we refactored the tool from the ground up. This CL adds it
back.

User-mode networking depends on SLiRP, which operates by defining a
fixed set of port mappings between the host and guest systems. The set
of possible ports are defined in the device_spec, associated with a
user-readable name (e.g. "ssh" is guest port 22). The host ports mapped
to those guest ports are then defined by the user at runtime on the
command line. Any ports defined in the device_spec that are not also
specified on the command line are assigned arbitrary ports dynamically
at runtime using the ffx::port_picker library.

The former "--tuntap" command-line flag has been replaced with the
"--net" flag, which can be "none", "user", "tap", or "auto".

The setup args template is then modified to populate the Qemu
command-line according to these mappings only when using the "--net
user" mode. The rest of the networking flags are unchanged (except to
recognize the modified flag values).

Original-Bug: fxbug.dev/88327
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/631087
Original-Revision: ed89abddaed43dd727503f234618e0d918c8781e
GitOrigin-RevId: 13cceda72718ee37ab04c62b2f2f0895b0dc77ae
Change-Id: Id42d37fd78c46ba94d384c53c555f84f63c62610