[roll] Roll fuchsia [ffx] Remove env flag from ffx emu
The ffx emu tool relies on a Product Bundle Manifest for the majority of
its configuration. While some of the fields in the manifest are
configurable through command lines, we wish to minimize the number of
these flags to keep the command line simple. Complicated configuration
should happen directly in the virtual device specification or the setup
flags template.

This CL removes the env flag from the emu tool. This flag was a holdover
from the previous version of the tool, was completely unused in the
current version, and is not the way we want users to modify the runtime
environment anyway.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632493
Original-Revision: 6afbd762306d372ebdcf4df99c5ed16edb14a8dd
GitOrigin-RevId: 381c17fccb261127522616524de2c3cf1fe1ae29
Change-Id: I4a8a3e9b8812344da80e0d578137488c0ea7e735