[roll] Roll fuchsia [tools][fuzz] Fix undercoat bug resulting in truncated QEMU log files
Previously, in order to detect a successful boot, undercoat relied on
monitoring the system's serial output via QEMU's stdout, looking for a
specific keyword. After detection, it then detaches from the subprocess.
Additionally, QEMU was also configured to copy its serial output to a
file.

When an unrecoverable error occurs in ClusterFuzz, the system debug logs
are fetched from the QEMU log file to aid in debugging. However, it was
noticed that these were being truncated at approximately 64KB.

The root cause was that, when QEMU is told to output serial to both
stdout and a file, a block on a write to one of those sinks blocks the
writes to the other as well. In this case, undercoat was only reading
the stdout until the boot completion keyword was detected, causing
writes after that to be buffered and then eventually block. Simply
continuing to read and discard from the pipe would also not be
sufficient, because the read end of the pipe is closed after the
`start_instance` command completes and undercoat exits.

To work around this, this CL switches QEMU to write solely to the log
file, which undercoat now watches during the boot process instead.

Tests are also added to verify the handling of logs of greater size than
default pipe buffer sizes, although the coverage is not perfect because
even in the end-to-end tests we do not invoke individual undercoat
commands as separate subprocesses.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632095
Original-Revision: d2065f850403aac11768c426a09eed15fd3aa0f8
GitOrigin-RevId: 49513d46679d2cc42cd2d45834e0ebcf6f136397
Change-Id: I52689821dd803acec3489931af8af4d196c5143c