[roll] Roll fuchsia [memfs] Change memfs C api implementation.
Implements the memfs C API in terms of the new C++ "Setup" object.

While this does not change the C API, it has an important semantic
change in that memfs_free_filesystem() will not remove the namespace
entries created my memfs_install_at() before shutdown. I believe failure
to do this is a bug so this change should not break any callers.

This removes all in-tree uses of memfs_uninstall_unsafe(). I believe
this function was working around the fact that the namespace entries
were not uninstalled and that memfs was not being shut down properly on
its thread.

Original-Bug: 91477
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/631642
Original-Revision: 658fe4da8dc056db8d459e5a7531d0caefe204c8
GitOrigin-RevId: ee6e41c8d0e61992adcb50ccb35329dc1a51ef2e
Change-Id: I664770c5cb9d622e3e201241ec9321a86bcad243