[roll] Roll fuchsia [fidl_driver] More LLCPP+Driver tests + fixes
This CL adds a broader set of tests for LLCPP+Drivers. These tests pass
when ASAN is disabled, but fail due to a use after free in teardown so
they are disabled. Still, it is useful to check them in to ensure the
code doesn't become stale.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632203
Original-Revision: 4cea6a47a872bd83ef76a796154efca1991084bc
GitOrigin-RevId: dad4eaa8f63c3f29190bc4c5239412a1630eec77
Change-Id: If4e955a366316507a9db0155bbd8bade5a98f0ec