[roll] Roll fuchsia [ffx] Add list filters for ancestors+descendants.
Introduces three new `--only` filters:

- `--only ancestors:foo`: Lists ancestors of a component named `foo`.
- `--only descendants:foo`: Lists descendants of a component named `foo`.
- `--only relatives:foo`: Lists both ancestors and descendants of `foo`.

This is not super useful for the `ffx component list` command, but
it is useful for the `ffx component graph` command, where it can be used
to significantly reduce the size of the graph for illustration.

The implementation accumulates the names of the ancestors for each
component in Component::parse.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/625602
Original-Revision: 9ad4bd38a3cfe61f423fcee6232a95ae3b7d4145
GitOrigin-RevId: cc04b612ac469c8780e6f4a0e98db508552bf8cb
Change-Id: Idb90ddb270ccf92f30871f4c657e3207db27f5b0