[roll] Roll fuchsia [audio_core][test] Tests for gain-ramp FIDL API
Add gain-control-test cases to cover error modes for the FIDL
API GainControl::SetGainWithRamp.

Tests verify that if the ramp's target gain_db is out of range
(above MAX_GAIN_DB, below MUTED_GAIN_DB, or is NAN), the
GainControl and AudioRenderer are both disconnected.

Other cases verify that if the ramp's duration is zero, or is
negative(!), the ramp call is redirected to a SetGain call.
This was already effectively occurring within the Gain class,
but new code in AudioRenderer makes this explicit and in turn
enables the Gain class to be more strict (subsequent CL).

fx test gain-control-test -- --gtest_filter=GainControlRampTest.*

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627878
Original-Revision: 3c030f03c818993670e1b2515062f6be7ccac86b
GitOrigin-RevId: fe42b49da5dc46dd93315c7b230101df1e606713
Change-Id: I03918cea10202b46656bf3002ef7d99ff88b6ab0