[roll] Roll fuchsia [fidl][rust] into_channel check message_interests.
When a message is sent which expects a response, a MessageResponse is
created to wait for it which holds a clone of the Arc<ChannelInner>,
which would prevent retrieving the raw channel through into_channel
while the message is outstanding. However, if the MessageResponse future
is dropped, it will release the Arc, but there will still be an
outstanding message which could come in later and be misinterpreted if
the channel is re-used for some other protocol.

To handle this case, check that the message_interests table is empty
before allowing the channel to be returned by into_channel.

Some tests in netstack3 relied on retrieving the channel with
into_channel and then checking that it was closed, however when the
channel is closed, message interests will end up in the Discard state
but never be dropped from the table. Toi cover this case, also allow
into_channel to succeed if the channel is closed, which is safe because
a closed channel can never receive additional messages, so won't have
any risk of getting still-pending messages later on.

Original-Bug: 68521
Test: fidldev test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/631027
Original-Revision: 947e549c9aa8b33b769c296309843106c62b16bc
GitOrigin-RevId: 56b7c4ef45da3d3abe1a74c65cdf1f4b3c91dd0e
Change-Id: Ic1b4dd799a3963ee95bd72f220fc582eefacc6c8