[roll] Roll fuchsia [fdf][compat] Connect protocol ops
Update the shim layers so the protocol ops are connected
correctly.

This CL looked at the DFv1 Driver Host for inspiration
on how these protocol ops were connected.

This CL updates the compat test so that it has a parent
driver and child driver, and the test now uses FIDL
to connect to the child driver, which connects to the
parent driver over Banjo.

Original-Bug: 90629
TEST: fx test compat-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/626582
Original-Revision: 31346190084c25204f354584fc6b7524bff7512a
GitOrigin-RevId: d28f2680b2bc4cd05b70186aa71a2b5de786caaa
Change-Id: I1107a7fadf7c8a9440e1dd19b04cdcc0ffe8479b