[roll] Roll fuchsia [fidlc] Rename Attributable to Element
This CL renames Attributable to Element and AttributePlacement to
Element::Kind. It also moves them from flat/attributes.h to flat_ast.h,
and splits out attribute_schema.{h,cc} to avoid circular dependencies.

This change brings the flat AST in line with the terminology from
"RFC-0083: FIDL Versioning", which defines a FIDL element as "a discrete
part of a FIDL library that affects the generated bindings." We had
already been using Attributable as a generic AST node in some places,
but it felt clunky, over-emphasizing the role of attributes. Now, the
focus is on elements (which happen to have attributes) rather than
attributable things (which happen to comprise all versioned AST nodes).

It is unlikely that these concepts will drift (i.e. new elements that
disallow attributes, or new attributable things that are not elements)
because elements can be versioned, and versioning is specified by
@available which is an attribute.

This also renames NamingContext::ElementKind to NamingContext::Kind to
avoid confusion, and removes the unused struct NamingContext::Element.

Test: fx test fidl-compiler
Original-Bug: 67858

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/631363
Original-Revision: 27d2d3a0683fb39b9d49eeaf44441b03cbcac338
GitOrigin-RevId: 442eebfeac50d7a32cb5b9b6a97345cc0585f7e0
Change-Id: I6a383a9578171421eb9b2658bea8eb1ad8dd4696