[roll] Roll fuchsia [connectivity] Move away from invaid use of cpp20::span for C++20
Our current cpp20::span polyfill accepts types with data() and size()
methods as ranges for its range constructor, but this is not the case
for the real std::span which cpp20::span falls back to when available.
Instead we now use the constructor that does take a pointer and size.

Test: Add --experimental_cxx_version=20 to args.gn and fx build
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632499
Original-Revision: 0d231a8efeaa865727f7687c260c1192c981e1ca
GitOrigin-RevId: c7bb1f352391c3c09e3b901d706627034532247e
Change-Id: I2b022b012375fbfa29934b245e9c03c473885cf0