[roll] Roll fuchsia [run-test-suite] Only output prefix when needed
In some infra builders the prefix is spammed for each line, which makes
the output very difficult to read. To resolve this issue, we keep track
of the last handle that wrote to the output stream, and only output the
prefix when the handle has changed.

Testing: fx test run-test-suite-lib-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621282
Original-Revision: 22ac7c8a9e39cdeb62f4a135318209a9c32a51c2
GitOrigin-RevId: 4e8174b9a4f4ca6e5eb8bea76484486d7ce8111e
Change-Id: I438b5ff721f82e3169801e745e10fedd3355c380