[roll] Roll cobalt Improve the comments and one name in the local aggregation proto.
Comments are added for every field.
Also moved one message so the ordering is consistent.
EventCodeAggregate is renamed to AggregationPeriodBucket. This better
reflects the contents of the message, and will make even more sense
when the EventCodesAggregateData is moved to the upcoming
SystemProfileAggregate. The naming also helps to differentiate from
EventCodesAggregateData and from all the other things that are called
FooAggregate (which typically use 'aggregate' as a variable name).

None of these changes affect the data, just the code.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/631062
Original-Revision: 6b84c2c99072df17e694934f31f804fa400b3608
GitOrigin-RevId: 2bedb306431ee3a8b099d06e7b9a802f1029e81a
Change-Id: I09891018f59173ba1d5de3761544083f3d9021e4