[roll] Roll fuchsia [sshutil] Mark ErrClosed as a connection error
This can happen when a keep-alive ping times out and the connection gets
closed in another goroutine.

This will cause testrunner to not consider a test to have failed if it
gets an ErrClosed, and to instead reconnect and retry running the test.

Original-Bug: 91584
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/633302
Original-Revision: 03b1031d028d2de8e730195538a89ae890f5790c
GitOrigin-RevId: eb000b1eb4307124a166e96a969b48cd8a4d9e00
Change-Id: I9a382a7f135ef90f566032909353c06b0948d38d