[roll] Roll fuchsia [tools][fuzz] Switch undercoat to C++ symbolizer
The Go symbolizer was removed from the base build in fxrev.dev/629945.
This CL switches over to using the C++ symbolizer, which also no longer
has a dependency on `llvm-symbolizer`.

Additionally, failing to be able to run the symbolizer was causing
fuzzer runs to hang, so this CL also improves handling of fuzzer output
processing to exit immediately with better errors messages in this sort
of circumstance.

Lastly, a new QEMU parameter was added to stop SeaBIOS from emitting
characters that interfered with terminal output and that was making it
slightly harder to read logs.

Original-Bug: 91640
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/633494
Original-Revision: 9e55d79a98098b0244310a51c5c2a3316528135d
GitOrigin-RevId: 5423a4be143ab39bdd5022a7adab3c6c6d88a199
Change-Id: I193fc208ee25aac6ba0f03162867762b405c8723