[roll] Roll infra/recipes [testing] Handle skipped and timed out tests
As of fxrev.dev/631582, testrunner will begin emitting "ABORT" and
"SKIP" statuses for tests so we should use that status to determine the
text to present for the test, whereas previously we had to compare the
test's runtime to the per-test timeout to determine if it timed out.
This lets us get rid of all the plumbing to pass the `per_test_timeout`
into the testing recipe module.

This actually needs to be landed *after* fxrev.dev/631582, otherwise we
will temporarily stop showing tests as timed-out because we don't emit
that status yet. It's safe to land in that order because recipes
currently assume a test is failed if it's status is anything other than
"PASS".

Original-Bug: 49266
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/631603
Original-Revision: ec4ebb130f0f884069b9394e8f32f8aab54c5e5a
GitOrigin-RevId: 50b113533c0f04bbde9adc137f3dabfa87582118
Change-Id: I513b3125b03c7bd2437175583cf7ea22ec4defa5