[roll] Roll fuchsia [fidlc] Enforce order in flat::Libraries
This CL changes flat::Libraries by updating documentation, adding
target_library(), and removing the Library* parameter from Unused(),
which only behaved correctly when given the final library. It also
removes TestLibrary::AddDependentLibrary.

This tightens the Libraries interface, making explicit the fact that
there is one main library compiled last. In particular, libraries must
be inserted in dependency order. The class was already used this way,
but its interface suggested it was simply a collection of libraries.

Tests were actually using Libraries in a slightly different way,
inserting all dependencies but not the final library as main.cc does.
This is now fixed by having TestLibrary::Compile do the insertion. As a
result, TestLibrary::AddDependentLibrary is no longer needed.

The motivation for this change is to move the "all libraries" Decl*
sorting to Libraries, which will be done in a follow up CL.

Test: fx test fidl-compiler
Original-Bug: 67858
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/633143
Original-Revision: 08582cee87b89fb43d17c57248e8e314d56badb4
GitOrigin-RevId: d84be3bf69f11940947f602cc47aafb44880d672
Change-Id: I1b58868fb4fb159417658055c2e11d1a192d0062