[roll] Roll fuchsia [kernel][vm] Avoid RefPtr copy
FindRegion has a redundant CopyPointer in order to work around
FindRegion not being a const method, but needing to potentially return
a const pointer. Copying, and then immediately destructing, this refptr
is a completely waste though.

Using a templated static that takes a self allows to effectively
template the constness of the method.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/633742
Original-Revision: 3830b4108883584c69f59cccf037f24f29df5379
GitOrigin-RevId: fbd2b389ff3b967f907d0c851532cc4d5a9ca220
Change-Id: I78fe064c245cb8e82bea61c7dba084aae51b006f