[roll] Roll fuchsia [Rust VFS] Allow seeking past end of file
Not allowing a zero position seek of a zero capacity file prevents
opening empty files, which broke the rust bootfs.

As Chris pointed out, per the standard we should actually allow seeking
an arbitrary amount past the end of the file. We're also meant to fill
the gap with zeroes if we ever write past the end of the file, and I'll
check that we do that correctly separately.

Documentation:
https://pubs.opengroup.org/onlinepubs/009696799/functions/lseek.html

Tested:
*Unit tests
*Created an empty file in '/boot', tried "cat david.txt" before and
 after

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/633110
Original-Revision: 9d16df773957f337962c796feb0c98ec0af091eb
GitOrigin-RevId: 3e54c676158b77b2b6853ad74a3a727a3f0e095b
Change-Id: Iaaf5df9960e74ec7ddab8a2a8077057cf4459a7c