[roll] Roll fuchsia [netstack3][ndp] Refactor IPv6 device addr config
Refactor the values in AddressEntry to constrain the presence of the
valid_until field to only those entries corresponding to
SLAAC-configured addresses. This makes it impossible to create an entry
for a manually-configured address with a non-empty valid_until value. It
also prepares the way for us to add additional information for
SLAAC-specified addresses in the future.

Original-Bug: 21428
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/623032
Original-Revision: 6d16e1192e5106363b89181e25327b08431a9114
GitOrigin-RevId: 3958f3db567a35f68e19011fce7799bc4b5de790
Change-Id: I5c77c44cafbb4d00e37df16a283a022231cb45a3