[roll] Roll fuchsia [component_manager] Replace `AbsoluteMoniker` constructor
Instead of creating an `AbsoluteMoniker`, then turning it into a
`PartialAbsoluteMoniker`, this change creates a `PartialAbsoluteMoniker`
directly, when used to obtain the instance id from a component. To do
so, `.to_partial()` is called on the existent `RelativeMoniker`
(required by `storage`) in `PartialAbsoluteMoniker::from_relative()`.

No behavioral change expected. This is refactoring only.

The reason for this change is to reduce the occurrences of
`AbsoluteMoniker`s if they are not needed, even in component manager.

Original-Bug: 87407
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632998
Original-Revision: 69fbf98f45b60b4ff85ed591d129d73429376695
GitOrigin-RevId: 61fe89c9e5d2e65f41196cb7c75517d25644ff09
Change-Id: I1c3e0fa79b28d064910388b5f58a60217690d5b7