[roll] Roll fuchsia [wlan][sae] Parameterize HMAC constructions
Thus far the kdf_sha256 implementation and the HMAC construction used in
our SAE support are hard-coded to use a SHA256 HMAC.  This is correct
for now, as we only support IKE Group 19 (IETF RFC 2409) which requires
SHA256.

This commit refactors the HMAC constructions to be parameterizable based
on the type of HMAC, which make these constructions more widely useful
and simplifies future support for other groups.

This commit also renames usages of the "h" function to "hkdf_extract",
as this function is precisely the function described in IETF RFC 5869,
and adds an implementation of the corresponding "hkdf_expand", which
will be used for WPA3 Hash-to-Element support.

Test: fx test wlan-sae-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632094
Original-Revision: 496f1b8c26535674c98471656a7d0fed3876adc9
GitOrigin-RevId: 543b25aff1ce5d7c668edab08fa0b25d5350e958
Change-Id: I3ffcf5c399bbbcbd7bcb416d795ccace3f730b1e