[roll] Roll fuchsia [fidl][cpp] No more inheriting from ClientBase
At this point, `WireClientImpl` only has async methods, so we can
replace it with `WireWeakAsyncClientImpl` and no longer inherit from
`ClientBase`. The more scalable pattern is to borrow a `ClientBase*`
and call methods on that.

Some tests were written with inheritance in mind. Those are also changed
to taking a `ClientBase*`.

It would be better to rename |ClientBase| to |ClientMessenger| or
similar since it is no longer a base class, but doing that later leads
to less intense churn.

Original-Fixed: 82085

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632682
Original-Revision: 2b63d803a674cf3bdd28447d88bb6fcabf51838f
GitOrigin-RevId: 64a1b4fafd584a474e2ee11fce0a9552bc814fde
Change-Id: Ib868c403f4ad192db025cf60f8c37de37b15f9cf