[roll] Roll fuchsia [ffx] Add --persist to ffx emu shutdown
One request we've received for the emulator is to have a mechanism for
persistent state for debugging, reuse, etc. This CL adds a flag to the
shutdown subcommand, "--persist", which instructs the tool not to remove
the state directory on shutdown.

This means after an instance is shut down, it will still show up in the
`ffx emu list` results. To help with this, we add a status indicator to
those results.

This also means we need to be pickier about the entries we return in the
list subcommand: the results should only include directory entries, and
they should contain an engine.json file; otherwise, they can't be valid
instances.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/634102
Original-Revision: c5ea6d2b71151cc68e6bd33aa4c3bf5f91821dcc
GitOrigin-RevId: 812261581f6a913bbe261207069c2cf66432aa7d
Change-Id: Ia9ca73f374c66277130d5dc01e3e0de39b0388dc