[roll] Roll fuchsia [paver] Fix std::span arguments for C++20
Our cpp20::span polyfill falls short, but the real std::span in C++20
doesn't work on temporaries to prevent lifetime issues. In this case the
data is copied out anyway so it's fine, but we still need to use lvalue
for span parameters for it to compile.

Test: Add --experimental_cxx_version=20 to args.gn and fx build
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/634212
Original-Revision: e9bc32111f1892684cbac4f66cf38f1e97a293c8
GitOrigin-RevId: cc4be1f5b5b06f74d09552d6b5fa4d1136b57053
Change-Id: I3df8d03b2038248dd9ac2c5ee980103f0ab80676