[roll] Roll fuchsia [test_manager] Move facet retrieval to a new mod
This moves facet retrieval and parsing to another module. This will be
useful as for debug data, we need to inspect the facets, then notify
the debug_data component of the moniker of the test suite (which
depends on the collection), then launch the suite.

Original-Bug: 77634
Testing: fx test test-manager-unittests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/631763
Original-Revision: 79ab30ad0802e5ce1b51e69276aa8c6cd960ec0d
GitOrigin-RevId: 7d153d9a3ae8d04ae3b4d51d4004d90e3cc20dde
Change-Id: I90dae415a3519f89bda0ecf4bf1b83902ffd9e8e