[roll] Roll fuchsia [kernel][x86] Add uninitialized
The inline asm block allows for either a register or memory output
operand, depending on what the compiler thinks is more optimal. The
compiler will first fill the memory location with the auto fill
pattern, then perform the asm block. This fill happens despite the
operand being marked output only, and not an input+output operand.

If the output constraints are tightened to only allow register output,
then the auto fill pattern is not performed. Perhaps the compiler
would also not auto fill when picking register if both reg and mem are
allowed, but I could not find a scenario where the compiler would ever
pick reg.

Marking the local variable __UNINITIALIZED prevents the compiler from
needing to auto fill, and retains the flexibility of allowing for both
register and memory outputs. This can most notably observed in
examining the objdump of spinlock code on x86 when saving the irq
state, as it otherwise ends up performing auto fills.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/626195
Original-Revision: 0990adb3de21fac2d1db9f233591d7bcd2d476a1
GitOrigin-RevId: d897e02f8f234e80e745c0c370b1ef1fdfab9aec
Change-Id: I1e0d5784867f1c023a98a2611104bd3058d39984