[roll] Roll fuchsia [fidl] Introduce breaking test case for @bindings_denylist bug
This is a test case that demonstrates erroneous behavior when processing
@bindings_denylist attributes. The next change in the chain should fix
this test, which will show up in the goldens.

The bug is as follows: the "child" definitions of a @bindings_denylisted
protocol or type should also be omitted from the resulting output. For
example, the method `DenyEachBinding.DenyGo`, nor any of its message
body types (request payload, response payload, response error, response
result union), should appear in the resuling go output, though this is
not the case today. Similarly, an anonymous layout defined with a
@bindings_denylisted declaration should be omitted from the output as
well.

Test: fidldev test all
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632494
Original-Revision: 04e7e950a471252a046c32f436aea46aa5a76202
GitOrigin-RevId: e26ea80d5ac6c4d1e789c0c38addf6a07b84e3ed
Change-Id: Ie6f44ee376aaf7efd180523032453f7de3fde884