[roll] Roll fuchsia [Rust VFS] Add a test for write after seek > size
Per the seek documentation, we should be able to seek past the end of a
file, write, and have the gap filled with zeroes. Adding this test since
we've just allowed seeking past end of file.

Documentation:
https://pubs.opengroup.org/onlinepubs/009696799/functions/lseek.html

Testing:
This is a test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/634443
Original-Revision: a73474a0bc0d5778d620db2e05ed75525dd1dd01
GitOrigin-RevId: aff6e60b4ac7221499b3114d1316f4c6ae9f3e6e
Change-Id: I126026f220aad8a551a85a7b8dbd57383d65d68d