[roll] Roll fuchsia [security][swd] Introduce common configuration for securty pkg tests
Much of the test configuration needed for security pkg tests
is not specific to a particular product. This change introduces
common configuration in a way that is sufficiently abstract to add
product-specific configuration as well. The code includes
placeholder component configuration data so that tests can be run
in-tree without additional files or template instantiations.

Original-Bug: 91099
Test: fx test fuchsia-pkg://fuchsia.com/core_access_ota_blob_as_executable#meta/core_access_ota_blob_as_executable.cm

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/629501
Original-Revision: 7a75cccdaa890a098c721809e184040c56115543
GitOrigin-RevId: 765456200baf387a4aa1eacfe501a888990e1f99
Change-Id: Idc9de7ba9c2b28bd079f0b2a5b82cd82c1f296b1