[roll] Roll fuchsia [fidl] Remove Struct.IsRequestOrResponse from fidlgen
To support named payloads, structs will no longer be allowed to "know"
whether or not they are used as a transactional message body, as it will
now be possible for a single struct to be used in both contexts.  This
change is the first in a series that will remove the
"is_request_or_reponse" JSON IR field, starting off by removing all
usages of the field in fidlgen.

Test: fidldev test all
Original-Bug: 88344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632581
Original-Revision: c4fef1dee6067f81dada09e748d9c8ff54da2e5c
GitOrigin-RevId: 25ec7653a279f72c87bf7c98533a0499cff1e516
Change-Id: Id67019eac0b05d5605c2aa08eff2a0a3e2bb44e1