[roll] Roll fuchsia [zxio] Let caller provide storage for dirents when enumerating
zxio provides a facility for enumerating over directory entries and
placing data into a zxio_dirent_t. This change lets the caller provide
the storage for the entries rather than returning a pointer to a zxio
internal pointer. This removes one data copy for the entries and
simplifies some of the lifecycle reasoning. The name field is typically
intended to go into a dirent struct, so this also lets the caller
specify where the name bytes should go in the zxio_dirent->name pointer.

With this change zxio is still buffering entries returned by the
filesystem server and not yet consumed by the client in an internal
buffer. A follow-up will let the caller also supply storage for this
buffer to remove the heap dependency inside zxio.

This API is used by https://fuchsia.googlesource.com/third_party/mesa
which is updated independently from fuchsia.git. To facilitate the API
transition this change adds a preprocessor define to zxio's public
configuration which mesa can use to select either the old or new API.
Once mesa is updated, we can remove this configuration.

Original-Bug: 91172
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/628538
Original-Revision: 1a07457919775b30c2bfacb0dd0a655af3b92d1c
GitOrigin-RevId: ec304102a80412df195ba4c1c78d9d2cd036c32c
Change-Id: I6c12d9a4b018a510cd5bdd1675113f0ddb4517f1