[roll] Roll fuchsia [netstack3] Support locally destined packets
Introduce a codepath for delivering packets to an IP address owned by
the stack. Route packets locally in the `ip::socket` and `ip` modules
when the destination address is assigned to the device we retrieve when
looking up a route for the packet.

Original-Bug: 42677

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/614647
Original-Revision: eb066083bc32be5b6460460195bb83a7b967a83a
GitOrigin-RevId: 5108eb4433dce1ad9daf693a93f123a7ed480521
Change-Id: Ic21946992f5b2da9bad9ea3269889e8a80715478