[roll] Roll fuchsia [scripts][fuzzing] Use sys.executable instead of 'python3.8'
When monitoring a fuzzer in the background, `fx fuzz` starts a python
subprocess. Previously it was doing this with an argument of
'python3.8', which didn't reliably find the prebuilt python. Using
sys.executable always finds the interpreter the script is using.

Original-Bug: 91827
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635121
Original-Revision: 0a2cf6f5d83817290afae5562cd3318d95068b14
GitOrigin-RevId: b9b0413bb6aae40adaaa3bc3ea8929dabce59284
Change-Id: I084f4c343671c21377ebb3dd795bd4b4eb692534