[roll] Roll fuchsia [component_manager] Hub returns error for instance not found
This CL replaces direct array accesses and `expect()`s for instances in
the `instance_map` with returning an error. If an instance is not found
in the `instance_map`, then a `ModelError` is returned, instead of a
panic. This emphasizes the responsibility of the caller to appropriately
handle `ModelErrors` produced from the hub event hooks.

The reason for this CL is prevent a panic and provide the opportunity
to handle a failure scenario, when an instance is not found in hub.rs.

Multiply: component-manager-tests: 1000
Multiply: hub_integration_test: 1000
Test: fx test component-manager-tests
      fx test hub_integration_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/633767
Original-Revision: 5f989f500c301921ee96912ecc8f75b17579e2e8
GitOrigin-RevId: cc9635ad6d77117867e5c24f06a34d0c30645fcb
Change-Id: I3ba6167464bbdf41f20b19225cc7238d708a8569