[roll] Roll fuchsia [fxfs] Move serialization into serialized_types.
This is an incremental change towards versioning but doesn't implement
any versioning changes yet. This change is intended as both an
illustration and a proof-of-concept but I expect further work to be
required before it's suitable to use.

This replaces calls to bincode serialize_into/deserialize_from with
equivalent calls to [Version] trait methods. The LSM tree code will be
tackled in a separate CL due to complications in the way it's
implemented (e.g. different types with the same layout are
serialized/deserialized).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/622842
Original-Revision: 0ae5926562140822c206dac856c9dbd9dcecb053
GitOrigin-RevId: 85e1c894986120c4bb000778587910c4062d27eb
Change-Id: I4841d27ca99da60e1926239eac3a74b6e1bfd027