[roll] Roll fuchsia [kernel][vm] Remove page_request_t from PageRequest
The only fields of page_request_t that are used PageProvider
implementations are the offset, len, type and queue_node. The first
three of these are replicated from PageRequest, and the queue_node can
be added to PageRequest. The rest of the page_request_t fields are not
used and so the entire page_request_t can be dropped out of
PageRequest.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/631344
Original-Revision: d0d416e2c29e86fd959524d4f5584a259581623d
GitOrigin-RevId: fbaf14d12ea2d1f89a522bbb13aebe9edd98dfd9
Change-Id: I623a2c4d8c9aa843dd1a5af6a90cadfd7f0f14da