[roll] Roll fuchsia [fdio][zxio] Use buffer+length strings in two path operations
This updates the logic for operations performed on two paths (rename and
link) to represent the string using an explicitly tracked length instead
of relying on null termination in both the zxio and fdio layers.

In zxio, the C interface is updated to accept a pointer to a buffer and
a length and the implementation is updated to use this length instead of
calling strlen() on the passed in value.

In fdio, the buffers used for these operations are allocated internally
on the stack. This updates the logic for allocating and manipulating the
path buffers to use fbl::StringBuffer which keeps track of the length
and uses std::string_view to pass buffer+length tuples inside fdio.

fbl::StringBuffer is also allocating and managing a null terminator, as
is the old code. This terminator isn't necessary and we may wish to
remove it later on.

Original-Bug: 91672
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/634644
Original-Revision: 99808ba955a236103064d8e2b7c8a3ef2eb2f856
GitOrigin-RevId: 83fba39cc3d2c42dfdb97dc39c9a265c646bb550
Change-Id: If6e15a7f891625f4e7767b1dc63f85cd6598a659