[roll] Roll fuchsia [system-update-checker] Grant fuchsia.space.Manager
When system-update-checker can run out of space when trying to update the
tuf metadata and fetch the update package to see if it should be installed.
Unfortunately when we transitioned over to CFv2, we forgot to include
passing fuchsia.space.Manager. We didn't catch this, because we didn't have
an integration test to make sure that we actually call Gc when we run out
of space.

This fixes the routing, and adds an integration test to make sure
everything works correctly.

Original-Fixed: 91884
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635370
Original-Revision: 6fca9eec3a1a738a80c9ca67b48216626fca3df7
GitOrigin-RevId: 1c7c84164b53f57cc87edad0376131b6d3960d75
Change-Id: I00347bd2f52290511438d0d6d2ed5890e3a1cb84