[roll] Roll fuchsia [debug_data] Ensure handle used for Wait is valid on Cancel
Move the inner state of the DataProcessor class to the data processor
thread so that the Zircon event handle it contains is still valid when
the Wait is cancelled. We need to do this as the underlying async
libraries assume the handles will be valid at the time a wait is
cancelled:

[00035.034491][1187][1312][processor] WARNING: ASSERT FAILED at (../../zircon/system/ulib/async-loop/loop.c:551): status == ZX_ERR_NOT_FOUND
[00035.034587][1187][1312][processor] WARNING: zx_port_cancel: status=-11

Testing: This issue only arises once we switch to using the new debug
data flow, in which the affected destructor is called. Verify by
running CQ with the change to use the new debug data flow.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/633765
Original-Revision: bad102d0edec720884ff4baae83ce86889d072f4
GitOrigin-RevId: a57a404cdee8d43bde3cf29d4151c4a9841b4b34
Change-Id: I6960c2df000ff95bd7724180930583f5399096cf