[roll] Roll fuchsia [network][tests] Fix fill send buffer flake
Before this change, to fill the send buffer of a socket on Linux, we did
non-blocking writes until EAGAIN was returned. This is flaky because the
peer may still acknowledge packets after we found EAGAIN on the sender,
which will free send buffer space.

Instead, we now check on the receiver side when its buffer is full, and
perform a final write once we know that no new data will be
acknowledged.

While I'm here, use an out parameter for the byte count and change the
return type to void. This allows the use of ASSERT_* macros.

Fix: 91464
Multiply: netstack_bsdsocket_c_api_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635034
Original-Revision: 168d9bd41daecfb8759704c5a8437e968198918f
GitOrigin-RevId: 93906828415cfb1b953d32bdc250da076195ea40
Change-Id: Id36296789713ec449f6ccb59e0244a17668d9620