[roll] Roll fuchsia [fidl] Enforce visibility for CPP bindings.
(Thanks to Aaron Wood for doing the hard work to isolate this).

We were not currently enforcing visibility constraints on the CPP
bindings for a FIDL build rule, unlike other languages. Since we tested
on other languages we didn't actually realize this.

Fortunately it looks like we only picked up two extra dependencies, one
in forensics which looks list just on omission in the visibility list,
and one in stash which I've created a bug to resolve.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632582
Original-Revision: 61bd0cdcf9b542e0eba25c40959f2bc169e0b846
GitOrigin-RevId: 76dd89d476524c8a56531fdf63b4d34a7a9859b9
Change-Id: Ia8a13bb5d6bdd67a93dd8e9ed4abd0d0003a6730