[roll] Roll fuchsia [physboot][handoff] Don't forget to report boot.timeline.phys-handoff
Fix a small bug in the predicate of a for loop which was causing the
kCount member of an enumeration to get skipped.  The net result of
this was that the `boot.timeline.phys-handoff` kcounter was always
staying with a value of 0, which was causing the
kernel-boot-benchmarks test/metric-collector to ignore the value
entirely.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/634741
Original-Revision: 4348bc810f1df5e1c846b1ea64f347e7d4a4b723
GitOrigin-RevId: e9902fa8078028106d78900bf03dfaee273f8d55
Change-Id: I10ca065c9c8f48b596b3c647b7ce3873244cd0bb