[roll] Roll fuchsia [zxio] Provide scratch buffer storage inline in zxio_dirent_iterator_t
The zxio directory iteration interface provides entries one at a time
for ease of use. The underlying transport operates in batches of entries
for efficiency, so something has to buffer the results until the caller
is ready to consume them. This change allocates storage for this
buffering inside the zxio_dirent_iterator_t structure.

Since this increases the size of the iterator considerably (from 64
bytes to slightly over 64 KiB) this also updates the fdio usage in
unistd.cc to allocate its iterator in the heap. Tests still generally
allocate this type on the stack. The other direct use of this interface
in the Mesa library is already allocating zxio_dirent_iterator_t in the
heap, so it is unchanged.

This also moves the type definition from zxio/zxio.h to zxio/types.h to
be more consistent with other type definitions.

Original-Bug: 91172
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/629583
Original-Revision: 1ac3debc7284d41d294aa6cce7a03686bb1a2d17
GitOrigin-RevId: da1c14beee0bf85bd210df254a7500c947883556
Change-Id: I25f80516a6c14a7b090f7e438ee90f94f04f3507