[roll] Roll fuchsia [pkg-resolver] BlobFetcher type alias -> wrapper struct
This no-op change is just to make it a bit easier to jump to the
fetching logic from the place where the fetch is enqueued. Hopefully it
makes pkg-resolver a little easier to navigate.

A pleasant side-effect is the doc comments get a bit nicer to read,
since they're a lot less abstract (since the queue's processing
behavior is known rather than being arbitrary)

This is a follow up to https://fxrev.dev/628587, which does the same for
the package fetch queue.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/630920
Original-Revision: 03ac6881bba90f8ac101019a47b9c49b77422686
GitOrigin-RevId: 43d713850700a6f150d2ff349541ec5502ba19be
Change-Id: I4a3bee1a1bf99b8208a167c9223c860af04b80bc