[roll] Roll fuchsia [run_test_suite] Custom run-scoped directory artifacts.
This CL supports retrieving a custom directory as an artifact from the
TestRun. The custom directory is scoped to the run itself rather than
any individual test case. This will be used in the future to support
including information about the complete state of a run to host-side
tools.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/619326
Original-Revision: 8a7ea13d9998c464865c010dc62c4c40adcfb85e
GitOrigin-RevId: 320acd7e46f7e7d2b54a1f322e170a51b2b5409c
Change-Id: I2c4a785cc7c770ce7c2620d43d1f8ea5db6aac12