[roll] Roll fuchsia [run-test-suite] Port early termination options to run-test-suite
Copies over some early termination options in ffx test to
run-test-suite. Since the tools use the same underlying library, this
simply exposes some options that are already available to the command
line.
This will simplify a fix to fx test that relies on these options. Since
fx test currently supports a fallback to run-test-suite, this is
necessary to ensure the options behave consistently when the fallback is
used and when it is not used.

Testing: Manually test behavior with `fx shell run-test-suite`
Original-Bug: 88478

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635375
Original-Revision: 5ad3c687df6c65b98d494186ac0681885fd25c2b
GitOrigin-RevId: a1861fc2a88f01a4b4617dfcec92b4447267a79e
Change-Id: I0dfeac37cccf7b0146e3a0ddb6e561a9decefc71