[roll] Roll fuchsia [brcmfmac] Handle suspend to prevent reboot crash
The driver would crash on every reboot because threads kept running
during shutdown. This happened because the driver was never notified of
the shutdown and so could not stop threads in an orderly fashion.

Fix this by making the Device class Suspendable and shutting everything
down when Suspend is called.

Original-Bug: 90694
Test: Manual rebooting and verifying no crashing
      fx test sim_device_test brcmfmac_sim_test brcmfmac_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635163
Original-Revision: 31feae88862e3f86e1bd06eb65542e2df1548da9
GitOrigin-RevId: e004ae55d234a78c14d104db3db40e6642233f09
Change-Id: I2489c8950be818e44bf56427dda2e9d97a4dd29f