[roll] Roll fuchsia [ffx] Manually add emu in user mode to ffx target
Before the rewrite, ffx emu would make a call to `ffx target add` to add
the emulator instance to the target list when the emulator was running
in user networking mode (i.e. SLiRP). This was needed because the IP
address included in any mDNS advertisements when running with user
networking would contain the internal IP address, which is invalid
outside of the emulation.

This CL reinstates that functionality in the new emu tool, adjusting for
changes to the `ffx target` code in the interim.

This introduces a bit of complexity when the emulator is running in
console/monitor mode, since the process is responsible for initiating
cleanup, but the shutdown code is also async. The "self" parameter from
the start code can't be shared between the threads launched for the two
commands, so we have to split out the parameters we want and pass them
explicitly and make the qemu_base::shutdown_emulator method an
associated function instead.

Original-Bug: 91642
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/634807
Original-Revision: 81ea3a57f11cd6b61c755a1b2d926096c87ba123
GitOrigin-RevId: c9517196a0cef0e120b98318440096b6f20849dd
Change-Id: Ie09b9dd28795e37ba8bbc8c6ae6db63d00e7ac17