[roll] Roll fuchsia [build] extend fastboot_flash image property
This CL documents the already-used fastboot_flash image property, which
is a list of partition names the image should be flashed to in a
standard flash operation.

In addition, this adds a special case that images can supply an empty
list in this property to indicate flash-adjacent images which should be
provided when flashing but aren't flashed directly to a partition, for
example unlock credentials and images used for bootstrapping.

Test: `fx flash` and `flash.sh` still work locally (Sherlock)
Test: `fx flash-remote` pulls unlock creds and bootstrap images
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/619061
Original-Revision: 0aa1bafa2ec712a0d22159f6f306f41e11951bdb
GitOrigin-RevId: 868e2c04b68f7599ec7fbfc1692a65040e4a3a49
Change-Id: Ia5df108723c5f1d5adabe283fcc07f22a5b46ad8