[roll] Roll fuchsia [iwlwifi] Support HT initial rate setting
The goal of this change is to set an initial 802.11n rate to the
firmware when association(to an AP which supports HT) is done.

The major part of this porting effort is the rate scale logic in
rateScaleMng.c, the rate selection part is not ported in this change.
Also some auxiliary functions are ported to make this effort more
smooth.

Original-Bug: 86522, 89011
Test: New test case added in wlan-softmac-device-test.cc.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585781
Original-Revision: a6cdafcb4a39afe76926254a22f4a223211af3f4
GitOrigin-RevId: de5dfcf9d24e4d9289c53aa6c1a278010f36d389
Change-Id: I396e269a4d51fcae1ef5541a5e5bca40dd8af5a8