[roll] Roll fuchsia [llcpp] Assert that server is channel when writing epitaphs
Reason::kClose should only happen when an epitaph is supposed to be
written, which is only supported on channels. This CL adds an assertion
to ensure that the server transport is a channel when a teardown with
Reason::kClose is taking place.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/628181
Original-Revision: 8ed706cabeb015182485a98a6e6fd623701d2c82
GitOrigin-RevId: 3dd6de54beb232cfbcfd235c414282818124481f
Change-Id: Ice9e7605ea555199ac56e2c203be815f4b86f115